### 2021-01-16 #2966 @whedon
[PRE REVIEW]: Agentpy: A package for agent-based modeling in Python
**Submitting author:** @JoelForamitti (<a href="http://orcid.org/0000-0002-4828-7288">Joël Foramitti</a>)
**Repository:** <a href="https://github.com/JoelForamitti/agentpy" target ="_blank">https://github.com/JoelForamitti/agentpy</a>
**Version:** v0.0.6
