+++ setup.cfg
@@ -15,3 +15,9 @@ install_requires =
     SALib
     IPython
     ipywidgets
+
+[options.extras_require]
+docs =
+    sphinx >= 3.2.1
+    nbsphinx >= 0.7.1
+    sphinx_rtd_theme >= 0.4.3
