### 2021-01-17 6719aa7
Update setup command
