+++ .readthedocs.yml
@@ -3,9 +3,9 @@ formats: all  # Include pdf & epub
 python:
   version: 3.7
   install:
+    - method: setuptools
+      path: .
     - method: pip
       path: .  # Install from current directory
       extra_requirements:
         - docs
-    - method: setuptools
-      path: .
