+++ .readthedocs.yml
@@ -5,7 +5,5 @@ python:
   install:
     - method: pip
       path: .
-    - method: pip
-      path: .  
       extra_requirements:
         - docs
