+++ .readthedocs.yml
@@ -7,3 +7,4 @@ python:
       path: .
       extra_requirements:
         - docs
+  system_packages: true
