+++ setup.cfg
@@ -1,11 +1,19 @@
 [metadata]
-version = attr: agentpy.__version__
+version = 0.0.7.dev
 
 [options]
 packages = find:
 python_requires = >=3.6
 setup_requires = pytest-runner
 tests_require = pytest, pytest-cov
+install_requires =
+    numpy >= 1.19
+    matplotlib >= 3.3.3
+    networkx >= 2.5
+    pandas >= 1.1.3
+    SALib >= 1.3.7
+    IPython >= 7.15.0
+    ipywidgets >= 7.5.1
 
 [options.extras_require]
 docs =
