### 2021-01-17 68b4c86
Update setup.cfg
