+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.0.7.dev
+version = attr: agentpy.__version__
 
 [options]
 packages = find:
@@ -17,6 +17,7 @@ install_requires =
 
 [options.extras_require]
 docs =
+    setuptools >= 46.4.0
     sphinx >= 3.2.1
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
