+++ .readthedocs.yml
@@ -4,6 +4,8 @@ python:
   version: 3.7
   install:
     - method: pip
+      path: setuptools >= 46.4.0
+    - method: pip
       path: .
       extra_requirements:
         - docs
