+++ setup.cfg
@@ -17,7 +17,6 @@ install_requires =
 
 [options.extras_require]
 docs =
-    setuptools >= 46.4.0
     sphinx >= 3.2.1
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
