+++ docs/requirements.txt
@@ -1,0 +1,1 @@
+setuptools >= 46.4.0
