+++ requirements.txt
@@ -1,10 +1,0 @@
-nbsphinx
-numpy
-scipy
-matplotlib
-networkx
-pandas
-SALib
-ipykernel
-IPython
-ipywidgets
