+++ .readthedocs.yml
@@ -2,11 +2,11 @@ version: 2
 formats: all
 python:
   version: 3.7
+  system_packages: true
   install:
-    - method: pip
-      path: setuptools >= 46.4.0
+    - requirements: docs/requirements.txt
     - method: pip
       path: .
       extra_requirements:
         - docs
-  system_packages: true
+
