+++ setup.cfg
@@ -20,10 +20,8 @@ docs =
     sphinx >= 3.2.1
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
-test =
+dev =
+    setuptools >= 46.4.0
     pytest >= 5.4.3
     coverage >= 5.3
-full =
-    seaborn >= 0.11.0
-    jupyter >= 1.0.0
 
