+++ docs/requirements.txt
@@ -1,1 +1,1 @@
-setuptools >= 46.4.0
+setuptools == 51.3.3
