+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = attr: agentpy.__version__
+version = 0.0.7.dev
 
 [options]
 packages = find:
@@ -21,7 +21,6 @@ docs =
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
 dev =
-    setuptools >= 46.4.0
     pytest >= 5.4.3
     coverage >= 5.3
 
