+++ .readthedocs.yml
@@ -4,7 +4,6 @@ python:
   version: 3.7
   system_packages: true
   install:
-    - requirements: docs/requirements.txt
     - method: pip
       path: .
       extra_requirements:
