### 2021-01-20 59145f6
Update version fetching
