+++ setup.cfg
@@ -7,6 +7,7 @@ python_requires = >=3.6
 setup_requires = pytest-runner
 tests_require = pytest, pytest-cov
 install_requires =
+    importlib-metadata ~= 1.0 ; python_version < "3.8"
     numpy >= 1.19
     matplotlib >= 3.3.3
     networkx >= 2.5
