+++ setup.cfg
@@ -22,6 +22,7 @@ docs =
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
 dev =
+    setuptools
     pytest >= 5.4.3
     coverage >= 5.3
 
