+++ setup.cfg
@@ -25,4 +25,12 @@ dev =
     setuptools
     pytest >= 5.4.3
     coverage >= 5.3
+min =
+    numpy == 1.19
+    matplotlib == 3.3.3
+    networkx == 2.5
+    pandas == 1.1.3
+    SALib == 1.3.7
+    IPython == 7.15.0
+    ipywidgets == 7.5.1
 
