+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.0.7.dev
+version = 0.0.7.dev0
 
 [options]
 packages = find:
@@ -25,12 +25,4 @@ dev =
     setuptools
     pytest >= 5.4.3
     coverage >= 5.3
-min =
-    numpy == 1.19
-    matplotlib == 3.3.3
-    networkx == 2.5
-    pandas == 1.1.3
-    SALib == 1.3.7
-    IPython == 7.15.0
-    ipywidgets == 7.5.1
 
