+++ .travis.yml
@@ -11,7 +11,7 @@ before_install:
   - pip install coverage
   - pip install codecov
 install:
-  - pip install .['min']
+  - pip install .
 script:
   - coverage run -m pytest
 after_success:
