### 2021-01-21 61b204f
Update minimum dependency tests
