+++ .travis.yml
@@ -12,6 +12,13 @@ before_install:
   - pip install codecov
 install:
   - pip install .
+  - pip install numpy==1.19
+  - pip install matplotlib==3.3.3
+  - pip install networkx==2.5
+  - pip install pandas==1.1.3
+  - pip install SALib==1.3.7
+  - pip install IPython==7.15.0
+  - pip install ipywidgets==7.5.1
 script:
   - coverage run -m pytest
 after_success:
