+++ .travis.yml
@@ -10,7 +10,7 @@ before_install:
   - pip install -U pytest
   - pip install coverage
   - pip install codecov
-install:
+install:  # Test minimum versions
   - pip install .
   - pip install numpy==1.19
   - pip install matplotlib==3.3.3
@@ -20,6 +20,6 @@ install:
   - pip install IPython==7.15.0
   - pip install ipywidgets==7.5.1
 script:
-  - coverage run -m pytest
+  - coverage run -m pytest  # run tests and assess coverage
 after_success:
-  - codecov # submit coverage
+  - codecov  # submit coverage
