+++ .readthedocs.yml
@@ -1,7 +1,11 @@
 version: 2
 formats: all
 python:
-  version: 3
-  install:
-    - requirements: requirements.txt
+  version: 3.7
   system_packages: true
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
+
