+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = attr: agentpy.__version__
+version = 0.0.7.dev0
 
 [options]
 packages = find:
@@ -7,11 +7,22 @@ python_requires = >=3.6
 setup_requires = pytest-runner
 tests_require = pytest, pytest-cov
 install_requires =
-    numpy
-    scipy
-    matplotlib
-    networkx
-    pandas
-    SALib
-    IPython
-    ipywidgets
+    importlib-metadata ~= 1.0 ; python_version < "3.8"
+    numpy >= 1.19
+    matplotlib >= 3.3.3
+    networkx >= 2.5
+    pandas >= 1.1.3
+    SALib >= 1.3.7
+    IPython >= 7.15.0
+    ipywidgets >= 7.5.1
+
+[options.extras_require]
+docs =
+    sphinx >= 3.2.1
+    nbsphinx >= 0.7.1
+    sphinx_rtd_theme >= 0.4.3
+dev =
+    setuptools
+    pytest >= 5.4.3
+    coverage >= 5.3
+
