### 2021-01-25 #2966 @danielskatz
@sbenthall  - you can ```@whedon assign me as editor``` whenever you are invited and agree, though I'll do it for you this time
