+++ setup.cfg
@@ -22,7 +22,10 @@ docs =
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
 dev =
-    setuptools
+    setuptools >= 51.3.3
     pytest >= 5.4.3
     coverage >= 5.3
+    sphinx >= 3.2.1
+    nbsphinx >= 0.7.1
+    sphinx_rtd_theme >= 0.4.3
 
