+++ docs/guide.rst
@@ -1,0 +1,12 @@
+.. currentmodule:: agentpy
+
+==========
+User guide
+==========
+
+.. :caption: Contents:
+.. toctree::
+   :maxdepth: 3
+
+   guide_random
+
