+++ docs/contribute.rst
@@ -1,0 +1,6 @@
+.. currentmodule:: agentpy
+
+==========
+Contribute
+==========
+
