+++ docs/guide.rst
@@ -1,12 +1,18 @@
 .. currentmodule:: agentpy
 
 ==========
-User guide
+User Guide
 ==========
+
+Welcome to the agentpy user guide.
+This section contains various articles to help with specific problems and applications.
+If you are interested to add a new article to this guide, please visit :doc:`contribute`.
+If you are looking for examples of complete models, take a look at :doc:`model_library`.
 
 .. :caption: Contents:
 .. toctree::
-   :maxdepth: 3
+   :caption: Contents
+   :maxdepth: 1
 
    guide_random
 
