+++ docs/about.rst
@@ -3,3 +3,17 @@
 =====
 About
 =====
+
+Agentpy has been created by Joël Foramitti and is
+available under the open-source `BSD 3-Clause <https://github.com/JoelForamitti/agentpy/blob/master/LICENSE>`_ license.
+Source files can be found on the `GitHub repository <https://github.com/joelforamitti/agentpy>`_.
+
+This project has benefited from an ERC Advanced Grant from the European Research Council (ERC)
+under the European Union's Horizon 2020 research and innovation programme (grant agreement n° 741087).
+
+Special thanks for their feedback and support go to Ivan Savin and Jeroen C.J.M van den Bergh.
+
+Parts of this package where created with Cookiecutter_ and the `audreyr/cookiecutter-pypackage`_ project template.
+
+.. _Cookiecutter: https://github.com/audreyr/cookiecutter
+.. _`audreyr/cookiecutter-pypackage`: https://github.com/audreyr/cookiecutter-pypackage
