+++ docs/guide.rst
@@ -6,7 +6,10 @@ User Guide
 
 Welcome to the agentpy user guide.
 This section contains various articles to help with specific problems and applications.
-If you are interested to add a new article to this guide, please visit :doc:`contribute`.
+Some of these articles are provided as interactive `Jupyter Notebooks <https://jupyter.org/>`_
+that can be downloaded and experimented with.
+
+If you are interested to add a new article to this guide, please visit :doc:`contributing`.
 If you are looking for examples of complete models, take a look at :doc:`model_library`.
 
 .. :caption: Contents:
