### 2021-01-26 299e9ce
Update test random shuffle
