+++ .travis.yml
@@ -13,7 +13,6 @@ before_install:
 install:
   - pip install .
   # Set minimum requirements
-  - pip install numpy==1.19
   - pip install matplotlib==3.3.3
   - pip install networkx==2.5
   - pip install pandas==1.1.3
