+++ agentpy/analysis.py
@@ -7,18 +7,18 @@ import matplotlib
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-import ipywidgets
-import IPython
 
 from SALib.analyze import sobol
 
 from .tools import make_list, param_tuples_to_salib
-from .objects import AgentList
 
 
-def sensitivity_sobol(output, param_ranges, measures=None, **kwargs):
+def sensitivity_sobol(output, param_ranges, measures=None,
+                      calc_second_order=False, **kwargs):
     """ Calculates Sobol Sensitivity Indices and adds them to the output,
-    using :func:`SALib.analyze.sobol.analyze`.
+    using :func:`SALib.analyze.sobol.analyze`. To work correctly,
+    data must be from an experiment with a parameter sample that was
+    generated by :func:`sample_saltelli`.
 
     Arguments:
         output (DataDict): The output of an experiment that was set to only
@@ -28,9 +28,10 @@ def sensitivity_sobol(output, param_ranges, measures=None, **kwargs):
             generation of the parameter sample with :func:`sample_saltelli`.
         measures (str or list of str, optional): The measures that should
             be used for the analysis. If none are passed, all are used.
+        calc_second_order (bool, optional): Whether to calculate second order
+            indices (default False). Value must be the same as the one used
+            to generate the parameter sample with :func:`sample_saltelli`.
         **kwargs: Will be forwarded to :func:`SALib.analyze.sobol.analyze`.
-            The kwarg ``calc_second_order`` must be the same as for
-            :func:`sample_saltelli`.
     """
 
     # STEP 1 - Convert param_ranges to SALib Format
@@ -39,35 +40,54 @@ def sensitivity_sobol(output, param_ranges, measures=None, **kwargs):
     param_ranges_salib = param_tuples_to_salib(param_ranges_tuples)
 
     # STEP 2 - Calculate Sobol Sensitivity Indices
-
     if measures is None:
         measures = output.measures.columns
-
     if isinstance(measures, str):
-        measures = make_list(measures)
-
-    dfs_si = []
-    dfs_si_conf = []
-    dfs_list = [dfs_si, dfs_si_conf]
+        measures = [measures]
+    p_keys = output._combine_pars(varied=True, fixed=False).keys()
+    dfs_list = [[] for _ in range(4 if calc_second_order else 2)]
 
     for measure in measures:
         y = np.array(output.measures[measure])
-        si = sobol.analyze(param_ranges_salib, y, **kwargs)
+        si = sobol.analyze(param_ranges_salib, y, calc_second_order, **kwargs)
 
-        # Make dataframes out of sensitivities
-        keys_list = [['S1', 'ST'], ['S1_conf', 'ST_conf']]
-        for dfs, keys in zip(dfs_list, keys_list):
-            s = {k: v for k, v in si.items() if k in keys}
+        # Make dataframes out of S1 and ST sensitivities
+        for dfs in dfs_list[0:2]:
+            s = {k: v for k, v in si.items() if k in ['S1', 'ST']}
             df = pd.DataFrame(s)
-            var_pars = output._combine_pars(varied=True, fixed=False)
-            df['parameter'] = var_pars.keys()
-            df['measure'] = measure
-            df = df.set_index(['measure', 'parameter'])
             dfs.append(df)
 
-    output['sensitivity'] = pd.concat(dfs_si)
-    output['sensitivity_conf'] = pd.concat(dfs_si_conf)
-    # TODO Second-Order Entries Missing
+        # Make dataframes out S2 sensitivities
+        if calc_second_order:
[…]
