### 2021-02-12 f7deb93
[docs] Fix jshtml control
