+++ docs/conf.py
@@ -90,3 +90,4 @@ html_theme = 'sphinx_rtd_theme'  # 'alabaster'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+html_css_files = ['css/custom.css']
