### 2021-02-26 #3065 @whedon
[REVIEW]: Agentpy: A package for agent-based modeling in Python
**Submitting author:** @JoelForamitti (<a href="http://orcid.org/0000-0002-4828-7288">Joël Foramitti</a>)
**Repository:** <a href="https://github.com/JoelForamitti/agentpy" target ="_blank">https://github.com/JoelForamitti/agentpy</a>
**Version:** v0.1.1
