+++ paper.bib
@@ -29,6 +29,7 @@
   volume    = {460},
   number    = {7256},
   pages     = {685--686},
+  doi       = {10.1038/460685a},
 }
 
 @Article{Castro2020,
@@ -39,7 +40,7 @@
   volume    = {11},
   number    = {4},
   pages     = {e647},
-  doi       = {https://doi.org/10.1002/wcc.647},
+  doi       = {10.1002/wcc.647},
 }
 
 @Article{Bianchi2015,
@@ -50,7 +51,7 @@
   volume    = {7},
   number    = {4},
   pages     = {284-306},
-  doi       = {https://doi.org/10.1002/wics.1356},
+  doi       = {10.1002/wics.1356},
 }
 
 @Article{Abar2017,
@@ -60,7 +61,7 @@
   year      = {2017},
   volume    = {24},
   pages     = {13 - 33},
-  doi       = {https://doi.org/10.1016/j.cosrev.2017.03.001},
+  doi       = {10.1016/j.cosrev.2017.03.001},
 }
 
 @Misc{Mesa2016,
@@ -75,5 +76,6 @@
   title={Managing business complexity: discovering strategic solutions with agent-based modeling and simulation},
   author={North, Michael J and Macal, Charles M},
   year={2007},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/acprof:oso/9780195172119.001.0001},
 }
