### 2021-03-08 a32f4ae
Added scipy requirement
