+++ .travis.yml
@@ -13,12 +13,12 @@ before_install:
 install:
   - pip install .
   # Set minimum requirements
-  - pip install matplotlib==3.3.3
-  - pip install networkx==2.5
-  - pip install pandas==1.1.3
-  - pip install SALib==1.3.7
-  - pip install IPython==7.15.0
-  - pip install ipywidgets==7.5.1
+  # - pip install matplotlib==3.3.3
+  # - pip install networkx==2.5
+  # - pip install pandas==1.1.3
+  # - pip install SALib==1.3.7
+  # - pip install IPython==7.15.0
+  # - pip install ipywidgets==7.5.1
 script:
   - coverage run -m pytest  # run tests and assess coverage
 after_success:
