+++ agentpy/grid.py
@@ -1,8 +1,13 @@
 """
 Agentpy Grid Module
-Content: Classes for spatial grids
+Content: Class for discrete spatial environments
 """
 
+# There is much room for optimization in this module. Contributions welcome! :)
+# TODO __init__: Add argument connect_borders
+# TODO _get_diamond: distance argument, exclude original agent, & limit dmax
+# TODO Method for distance between agents
+# TODO items: reverse argument
 
 import itertools
 import numpy as np
@@ -14,29 +19,28 @@ from .lists import AgentList
 
 
 class Grid(ApEnv):
-    """ Environment that contains agents with a spatial topology.
+    """ Environment that contains agents with a discrete spatial topology.
     Every location consists of an :class:`AgentList` that can hold
     zero, one, or more agents.
     To add new grid environments to a model, use :func:`Model.add_grid`.
+    For a continuous spatial topology, use :class:`Space`.
 
-    This class can be used as a parent class for custom network types.
+    This class can be used as a parent class for custom grid types.
     All agentpy model objects call the method :func:`setup` after creation,
     and can access class attributes like dictionary items.
     See :class:`Environment` for general properties of all environments.
 
     Arguments:
-        model(Model): The model instance.
-        shape(int or tuple of int): Size of the grid.
-            If an integer is given, this value is taken as both the
-            height and width for a two-dimensional grid.
-            If a tuple is given, the length of the tuple defines
-            the number of dimensions, and the values in the tuple
-            define the length of each dimension.
+        model (Model): The model instance.
+        shape (tuple of int): Size of the grid.
+            The length of the tuple defines the number of dimensions,
+            and the values in the tuple define the length of each dimension.
         **kwargs: Will be forwarded to :func:`Grid.setup`.
 
     Attributes:
-        grid(list of lists): Matrix of :class:`AgentList`.
-        shape(tuple of int): Length of each grid dimension.
+        grid (list of lists): Matrix of :class:`AgentList`.
+        shape (tuple of int): Length of each grid dimension.
+        dim (int): Number of dimensions.
     """
 
     def __init__(self, model, shape, **kwargs):
@@ -44,8 +48,6 @@ class Grid(ApEnv):
         super().__init__(model)
 
         self._topology = 'grid'
-        if isinstance(shape, int):
-            shape = (shape, shape)
         self._grid = make_matrix(shape, AgentList)
         self._agent_dict = {}
         self._shape = shape
@@ -60,20 +62,27 @@ class Grid(ApEnv):
     def shape(self):
         return self._shape
 
+    @property
+    def dim(self):
+        return len(self._shape)
+
     def add_agents(self, agents=1, agent_class=Agent, positions=None,  # noqa
-                   random=False, **kwargs):
-        """ Adds agents to the grid environment.
+                   random=False, generator=None, **kwargs):
+        """ Adds agents to the grid environment, and returns new agents.
         See :func:`Environment.add_agents` for standard arguments.
         Additional arguments are listed below.
 
         Arguments:
-            positions(list of tuples, optional): The positions of the added
+            positions(list of tuple, optional): The positions of the added
                 agents. List must have the same length as number of agents
                 to be added, and each entry must be a tuple with coordinates.
                 If none is passed, agents will fill up the grid systematically.
             random(bool, optional):
                 If no positions are passed, agents will be placed in random
                 locations instead of systematic filling (default False).
+            generator (numpy.random.Generator, optional):
+                Random number generator that is used if 'random' is True.
+                If none is passed, :obj:`Model.random` is used.
         """
 
[…]
