### 2021-03-11 35f59bb
[fix] test_random
