### 2021-03-11 24859f5
[fix] Scipy for Python 3.6
