+++ setup.cfg
@@ -15,7 +15,7 @@ install_requires =
     SALib >= 1.3.7
     IPython >= 7.15.0
     ipywidgets >= 7.5.1
-    scipy >= 1.6.0
+    scipy >= 1.5.2
 
 [options.extras_require]
 docs =
