### 2021-03-24 7895174
[new] Ipysimulate Methods
