### 2021-03-25 bc15296
Iterator features
