+++ tests/test_lists.py
@@ -12,7 +12,7 @@ def test_repr():
     assert model.objects.__repr__() == "ObjList [2 objects]"
     l1 = model.agents.id
     l2 = l1 + 1
-    assert l1.__repr__() == "AttrList of attribute 'id': [1]"
+    assert l1.__repr__() == "AttrList of 'id': [1]"
     assert l2.__repr__() == "AttrList: [2]"
 
 
@@ -24,7 +24,7 @@ def test_attr_calls():
     assert list(model.agents.x) == [1, 1]
     assert list(model.agents.f()) == [2, 2]
     with pytest.raises(AttributeError):
-        assert model.agents.y
+        assert list(model.agents.y)  # Convert to list to call attribute
     with pytest.raises(TypeError):
         assert model.agents.x()  # noqa
 
