### 2021-03-26 67cc917
[new] alive property
