### 2021-03-26 65e9f08
[fix] Unnecessary set_var_ignore
