### 2021-03-26 c8a685f
[fix] Empty neighbor return
