### 2021-03-26 470bdf5
[fix] check alive for remove_agents
