### 2021-03-26 061b7f6
[docs] Property descriptions
