+++ docs/about.rst
@@ -8,12 +8,26 @@ Agentpy has been created by Joël Foramitti and is
 available under the open-source `BSD 3-Clause <https://github.com/JoelForamitti/agentpy/blob/master/LICENSE>`_ license.
 Source files can be found on the `GitHub repository <https://github.com/joelforamitti/agentpy>`_.
 
-This project has benefited from an ERC Advanced Grant from the European Research Council (ERC)
-under the European Union's Horizon 2020 research and innovation programme (grant agreement n° 741087).
+This package would not have been possible without the great work and
+support from the Python and open-source community.
+Particular thanks for making their work and knowledge publicly available go
+to the following users:
 
-Special thanks for their feedback and support go to Ivan Savin and Jeroen C.J.M van den Bergh.
+- Thierry Lathuille https://stackoverflow.com/users/550094/
+- Audrey Feldroy https://github.com/audreyfeldroy/
 
-Parts of this package where created with Cookiecutter_ and the `audreyr/cookiecutter-pypackage`_ project template.
+Special thanks for feedback and support:
+
+- Ivan Savin
+- Jeroen C.J.M van den Bergh
+
+Parts of this package where created with Cookiecutter_
+and the `audreyr/cookiecutter-pypackage`_ project template.
+
+This project has further benefited from an ERC Advanced Grant
+from the European Research Council (ERC)
+under the European Union's Horizon 2020 research and innovation programme
+(grant agreement n° 741087).
 
 .. _Cookiecutter: https://github.com/audreyr/cookiecutter
 .. _`audreyr/cookiecutter-pypackage`: https://github.com/audreyr/cookiecutter-pypackage
