### 2021-04-06 #3065 @JoelForamitti
@jamesdamillington, thank you for your comment! From other publications in JOSS (see e.g. [this recent one](https://joss.theoj.org/papers/10.21105/joss.03097)) I had the impression that it is allowed to explain the need without an explicit section title of that name. @sbenthall, could you confirm or refute this?
