### 2021-04-09 #3065 @jamesdamillington
@JoelForamitti @sbenthall I've completed my review at [JOSS Review (JM)](https://github.com/JoelForamitti/agentpy/issues/13) and conditionally accept the submission depending on the issues highlighted there. Looks good! Any queries on my comments, just let me know.
