### 2021-04-16 #3065 @xuanxu
Yes, the `Statement of need` is a required section and we just [recently](https://github.com/openjournals/joss/pull/883) [changed](https://github.com/openjournals/joss/pull/862) the documentation to better refelct this and avoid exceptions like the mentioned paper.
