+++ agentpy/analysis.py
@@ -13,85 +13,6 @@ from SALib.analyze import sobol
 from .tools import make_list, param_tuples_to_salib
 
 
-def _sobol_set_df_index(df, p_keys, measure):
-    df['parameter'] = p_keys
-    df['measure'] = measure
-    df.set_index(['measure', 'parameter'], inplace=True)
-
-
-def sensitivity_sobol(output, param_ranges, measures=None,
-                      calc_second_order=False, **kwargs):
-    """ Calculates Sobol Sensitivity Indices and adds them to the output,
-    using :func:`SALib.analyze.sobol.analyze`. To work correctly,
-    data must be from an experiment with a parameter sample that was
-    generated by :func:`sample_saltelli`.
-
-    Arguments:
-        output (DataDict): The output of an experiment that was set to only
-            one iteration (default) and used a parameter sample that was
-            generated with :func:`sample_saltelli`.
-        param_ranges (dict): The same dictionary that was used for the
-            generation of the parameter sample with :func:`sample_saltelli`.
-        measures (str or list of str, optional): The measures that should
-            be used for the analysis. If none are passed, all are used.
-        calc_second_order (bool, optional): Whether to calculate second order
-            indices (default False). Value must be the same as the one used
-            to generate the parameter sample with :func:`sample_saltelli`.
-        **kwargs: Will be forwarded to :func:`SALib.analyze.sobol.analyze`.
-    """
-
-    # STEP 1 - Convert param_ranges to SALib Format
-    param_ranges_tuples = {k: v for k, v in param_ranges.items()
-                           if isinstance(v, tuple)}
-    param_ranges_salib = param_tuples_to_salib(param_ranges_tuples)
-
-    # STEP 2 - Calculate Sobol Sensitivity Indices
-    if measures is None:
-        measures = output.measures.columns
-    if isinstance(measures, str):
-        measures = [measures]
-    p_keys = output._combine_pars(varied=True, fixed=False).keys()
-    dfs_list = [[] for _ in range(4 if calc_second_order else 2)]
-
-    for measure in measures:
-        y = np.array(output.measures[measure])
-        si = sobol.analyze(param_ranges_salib, y, calc_second_order, **kwargs)
-
-        # Make dataframes out of S1 and ST sensitivities
-        keyss = [['S1', 'ST'], ['S1_conf', 'ST_conf']]
-        for keys, dfs in zip(keyss, dfs_list[0:2]):
-            s = {k[0:2]: v for k, v in si.items() if k in keys}
-            df = pd.DataFrame(s)
-            _sobol_set_df_index(df, p_keys, measure)
-            dfs.append(df)
-
-        # Make dataframes out S2 sensitivities
-        if calc_second_order:
-            for key, dfs in zip(['S2', 'S2_conf'], dfs_list[2:4]):
-                df = pd.DataFrame(si[key])
-                _sobol_set_df_index(df, p_keys, measure)
-                dfs.append(df)
-
-    # Combine dataframes for each measure
-    output['sensitivity'] = pd.concat(dfs_list[0])
-    output['sensitivity_conf'] = pd.concat(dfs_list[1])
-
-    if calc_second_order:
-
-        # Add Second-Order to Output
-        dfs_si = [output['sensitivity'], pd.concat(dfs_list[2])]
-        dfs_si_conf = [output['sensitivity_conf'], pd.concat(dfs_list[3])]
-        output['sensitivity'] = pd.concat(dfs_si, axis=1)
-        output['sensitivity_conf'] = pd.concat(dfs_si_conf, axis=1)
-
-        # Create Multi-Index for Columns
-        arrays = [["S1", "ST"] + ["S2"] * len(p_keys), [""] * 2 + list(p_keys)]
[…]
