+++ agentpy/experiment.py
@@ -7,9 +7,13 @@ import pandas as pd
 import ipywidgets
 import IPython
 
+from os import sys
+
+from .version import __version__
 from datetime import datetime, timedelta
 from .tools import make_list
-from .output import DataDict
+from .datadict import DataDict
+from .sample import Sample
 
 
 class Experiment:
@@ -19,57 +23,70 @@ class Experiment:
     simulations and :func:`Experiment.interactive` for interactive output.
 
     Arguments:
-        model_class(type): The model class type that the experiment should use.
-        parameters(dict or list of dict, optional):
-            Parameter dictionary or sample (default None).
-        name(str, optional): Name of the experiment (default model.name).
-        scenarios(str or list, optional): Experiment scenarios (default None).
-        iterations(int, optional): Experiment repetitions (default 1).
-        record(bool, optional):
-            Whether to keep the record of dynamic variables (default False).
-            Note that this does not affect evaluation measures.
-        **kwargs: Will be forwarded to the creation of every model instance
-            during the experiment.
+        model (type):
+            The :class:`Model` type to perform the experiment with.
+        sample (dict or list of dict or Sample, optional):
+            Parameter combination(s) to test in the experiment (default None).
+        iterations (int, optional):
+            How often to repeat every parameter combination (default 1).
+        random (bool, optional):
+            Choose random seeds for every new iteration (default False).
+            The seed for the random number generator will be taken from the
+            model's current parameter combination.
+        record (bool, optional):
+            Keep the record of dynamic variables (default False).
+        **kwargs:
+            Will be forwarded to all model instances created by the experiment.
 
     Attributes:
         output(DataDict): Recorded experiment data
     """
 
-    def __init__(self, model_class, parameters=None, name=None, scenarios=None,
-                 iterations=1, record=False, **kwargs):
+    def __init__(self, model_class, sample=None, iterations=1,
+                 random=False, record=False, **kwargs):
 
         self.model = model_class
         self.output = DataDict()
         self.iterations = iterations
         self.record = record
         self._model_kwargs = kwargs
+        self._random = False
+        self.name = model_class.__name__
 
-        if name:
-            self.name = name
+        # TODO Improve
+        if isinstance(sample, Sample):
+            self.sample = list(sample)
+            sample_type = sample._type
         else:
-            self.name = model_class.__name__
-
-        # Transform input into iterable lists if only a single value is given
-        # keep_none assures that make_list(None) returns iterable [None]
-        self.scenarios = make_list(scenarios, keep_none=True)
-        self.parameters = make_list(parameters, keep_none=True)
-        self._parameters_to_output()  # Record parameters
-
-        # Log
-        self.output.log = {'name': self.name,
-                           'model_type': model_class.__name__,
-                           'time_stamp': str(datetime.now()),
-                           'iterations': iterations}
-        if scenarios:
-            self.output.log['scenarios'] = scenarios
+            self.sample = make_list(sample, keep_none=True)
+            sample_type = 'custom'
 
         # Prepare runs
-        self.parameters_per_run = self.parameters * self.iterations
-        self.number_of_runs = len(self.parameters_per_run)
+        self.run_ids = [(sample_id, iteration)
+                        for sample_id in range(len(self.sample))
+                        for iteration in range(iterations)]
+        self.n_runs = len(self.run_ids)
+
+        # Prepare output
+        self.output.log = {
+            'model_type': model_class.__name__,
+            'time_stamp': str(datetime.now()),
+            'agentpy_version': __version__,
+            'python_version': sys.version[:5],
+            'multi_run': True,
[…]
