### 2021-04-30 6607f47
New structure (numpy.rec)
