+++ agentpy/grid.py
@@ -3,279 +3,391 @@ Agentpy Grid Module
 Content: Class for discrete spatial environments
 """
 
-# There is much room for optimization in this module. Contributions welcome! :)
-# TODO __init__: Add argument connect_borders
-# TODO _get_diamond: distance argument, exclude original agent, & limit dmax
-# TODO Method for distance between agents
-# TODO items: reverse argument
+# TODO Agent Fields - How to combine with track_empty?
 
 import itertools
 import numpy as np
 import random as rd
 import collections.abc as abc
-from .objects import ApEnv, Agent, Location
-from .tools import make_list, make_matrix
-from .lists import AgentList, ObjList, LocList
+import numpy.lib.recfunctions as rfs
+from .objects import Spatial
+from .tools import make_list, make_matrix, AgentpyError, ListDict
+from .sequences import AgentSet, AgentIter, AgentList
 
 
-class Grid(ApEnv):
-    """ Environment that contains agents with a discrete spatial topology.
-    Every location consists of an :class:`AgentList` that can hold
-    zero, one, or more agents.
-    To add new grid environments to a model, use :func:`Model.add_grid`.
-    For a continuous spatial topology, use :class:`Space`.
+class _IterArea:
+    """ Iteratable object that takes either a numpy matrix or an iterable
+    as an input. If the object is an ndarray, it is flattened and iterated
+    over the contents of each element chained together. Otherwise, it is
+    simply iterated over the object.
 
+    Arguments:
+        area: Area of sets of elements.
+        exclude: Element to exclude. Assumes that element is in area.
+    """
+
+    def __init__(self, area, exclude=None):
+        self.area = area
+        self.exclude = exclude
+
+    def __len__(self):
+        if isinstance(self.area, np.ndarray):
+            len_ = sum([len(s) for s in self.area.flat])
+        else:
+            len_ = len(self.area)
+        if self.exclude:
+            len_ -= 1  # Assumes that exclude is in Area
+        return len_
+
+    def __iter__(self):
+        if self.exclude:
+            if isinstance(self.area, np.ndarray):
+                return itertools.filterfalse(
+                    lambda x: x is self.exclude,
+                    itertools.chain.from_iterable(self.area.flat)
+                )
+            else:
+                return itertools.filterfalse(
+                    lambda x: x is self.exclude, area)
+        else:
+            if isinstance(self.area, np.ndarray):
+                return itertools.chain.from_iterable(self.area.flat)
+            else:
+                return iter(area)
+
+
+class GridIter(AgentIter):
+    """ Iterator over objects in :class:`Grid` that supports slicing.
+
+    Examples:
+
+         Create a model with a 10 by 10 grid
+         with one agent in each position::
+
+            model = ap.Model()
+            agents = ap.AgentList(model, 100)
+            grid = ap.Grid(model, (10, 10))
+            grid.add_agents(agents)
+
+        The following returns an iterator over the agents in all position::
+
+            >>> grid.agents
+            GridIter (100 objects)
+
+        The following returns an iterator over the agents
+        in the top-left quarter of the grid::
+
+            >>> grid.agents[0:5, 0:5]
+            GridIter (25 objects)
+    """
+
+    def __init__(self, iter_, items):
+        super().__init__(iter_)
+        object.__setattr__(self, '_items', items)
+
+    def __getitem__(self, item):
+        sub_area = self._items[item]
+        return GridIter(_IterArea(sub_area), sub_area)
+
+
+class Grid(Spatial):
+    """ Environment that contains agents with a discrete spatial topology,
+    supporting both multiple agents per cell and multiple fields per grid.
     This class can be used as a parent class for custom grid types.
     All agentpy model objects call the method :func:`setup` after creation,
[…]
