### 2021-04-30 95525bd
New structures (AgentGroup)
