+++ agentpy/lists.py
@@ -1,377 +1,0 @@
-"""
-Agentpy Lists Module
-Content: Lists for objects, environments, and agents
-"""
-
-import numpy as np
-
-
-class AttrList:
-    """ Iterator of attributes from an :class:`AgentList`.
-
-    Calls are forwarded to each entry and return a list of return values.
-    Boolean operators are applied to each entry and return a list of bools.
-    Arithmetic operators are applied to each entry and return a new list.
-    Item assignments are forwarded to the object in the original list.
-    See :class:`AgentList` for examples.
-    """
-
-    def __init__(self, source, attr=None):
-        self.source = source
-        self._iter_source = None
-        self.attr = attr
-
-    def _repr_list(self):
-        rep = "["
-        first = True
-        for el in self:
-            if first:
-                first = False
-            else:
-                rep += ', '
-            if isinstance(el, AttrList):
-                rep += el.__repr__(short=True)
-            else:
-                rep += el.__repr__()
-        return rep + ']'
-
-    # TODO IN DEVELOPMENT
-    def flatten(self, levels=1, sum=False, empty=np.nan):
-        if isinstance(list_[0], AttrList):
-            return [self.flatten(sublist)
-                    for sublist in list_]
-        else:
-            return #[*self.flatten(sublist)
-                   # for sublist in list_]
-        #else:
-        #    return [func(i, *args, **kwargs) for i in grid]
-
-    def __repr__(self, short=False):
-        rep = ""
-        if not short:
-            rep += "AttrList"
-            if self.attr:
-                rep += f"of '{self.attr}'"
-            rep += ": "
-        return rep + self._repr_list()
-
-    def __iter__(self):
-        """ Iterate through source list based on attribute. """
-        if self.attr:
-            a = self.attr
-            return iter([getattr(o, a) for o in self.source])
-            #for el in self.source:
-            #    yield getattr(el, a)
-        else:
-            return iter(self.source)
-            #for el in self.source:
-            #    yield el  # return iter(self.source)
-
-    def __getitem__(self, key):
-        """ Get item from source list. """
-        return getattr(self.source[key], self.attr)
-
-    def __setitem__(self, key, value):
-        """ Set item to source list. """
-        setattr(self.source[key], self.attr, value)
-
-    def __call__(self, *args, **kwargs):
-        return [func_obj(*args, **kwargs) for func_obj in self]
-
-    def __eq__(self, other):
-        # TODO doesn't work with second attrlist
-        return [obj == other for obj in self]
-
-    def __ne__(self, other):
-        return [obj != other for obj in self]
-
-    def __lt__(self, other):
-        return [obj < other for obj in self]
-
-    def __le__(self, other):
-        return [obj <= other for obj in self]
-
-    def __gt__(self, other):
-        return [obj > other for obj in self]
-
-    def __ge__(self, other):
-        return [obj >= other for obj in self]
-
-    def __add__(self, v):
-        if isinstance(v, AttrList):
-            return AttrList([x + y for x, y in zip(self, v)])
-        else:
-            return AttrList([x + v for x in self])
-
-    def __sub__(self, v):
-        if isinstance(v, AttrList):
-            return AttrList([x - y for x, y in zip(self, v)])
-        else:
-            return AttrList([x - v for x in self])
-
-    def __mul__(self, v):
-        if isinstance(v, AttrList):
-            return AttrList([x * y for x, y in zip(self, v)])
-        else:
-            return AttrList([x * v for x in self])
-
-    def __truediv__(self, v):
-        if isinstance(v, AttrList):
-            return AttrList([x / y for x, y in zip(self, v)])
-        else:
-            return AttrList([x / v for x in self])
-
[…]
