+++ agentpy/sequences.py
@@ -1,0 +1,466 @@
+"""
+Agentpy Lists Module
+Content: Lists for objects, environments, and agents
+"""
+
+# TODO Seperate AgentIter & AgentGroupIter
+
+import itertools
+import agentpy as ap
+import numpy as np
+from .tools import AgentpyError
+from collections.abc import Sequence as PySequence
+
+
+class Sequence:
+    """ Base class for agenpty sequences. """
+
+    def __repr__(self):
+        s = 's' if len(self) != 1 else ''
+        return f"{type(self).__name__} ({len(self)} object{s})"
+
+    def __getattr__(self, name):
+        # TODO This breaks numpy conversion because of __array_struct__ lookup
+        """ Return callable list of attributes """
+        if name[0] == '_':  # Allows for numpy conversion
+            super().__getattr__(name)  #raise AttributeError('test')
+        else:
+            return AttrIter(self, attr=name)
+
+    def _set(self, key, value):
+        object.__setattr__(self, key, value)
+
+    @staticmethod
+    def _obj_gen(model, n, cls, *args, **kwargs):
+        if cls is None:
+            cls = ap.Agent
+        for _ in range(n):
+            yield cls(model, *args, **kwargs)
+
+
+# Attribute List ------------------------------------------------------------ #
+
+class AttrIter(Sequence, PySequence):
+    """ Iterator over an attribute of objects in a sequence.
+    Length, items access, and representation work like with a normal list.
+    Calls are forwarded to each entry and return a list of return values.
+    Boolean operators are applied to each entry and return a list of bools.
+    Arithmetic operators are applied to each entry and return a new list.
+    If applied to another `AttrList`, the first entry of the first list
+    will be matched with the first entry of the second list, and so on.
+    Else, the same value will be applied to each entry of the list.
+    See :class:`AgentList` for examples.
+    """
+
+    def __init__(self, source, attr=None):
+        self.source = source
+        self.attr = attr
+
+    def __repr__(self):
+        return repr(list(self))
+
+    @staticmethod
+    def _iter_attr(a, s):
+        for o in s:
+            yield getattr(o, a)
+
+    def __iter__(self):
+        """ Iterate through source list based on attribute. """
+        if self.attr:
+            return self._iter_attr(self.attr, self.source)
+        else:
+            return iter(self.source)
+
+    def __len__(self):
+        return len(self.source)
+
+    def __getitem__(self, key):
+        """ Get item from source list. """
+        return getattr(self.source[key], self.attr)
+
+    def __setitem__(self, key, value):
+        """ Set item to source list. """
+        setattr(self.source[key], self.attr, value)
+
+    def __call__(self, *args, **kwargs):
+        return [func_obj(*args, **kwargs) for func_obj in self]
+
+    def __eq__(self, other):
+        return [obj == other for obj in self]
+
+    def __ne__(self, other):
+        return [obj != other for obj in self]
+
+    def __lt__(self, other):
+        return [obj < other for obj in self]
+
+    def __le__(self, other):
+        return [obj <= other for obj in self]
+
+    def __gt__(self, other):
+        return [obj > other for obj in self]
+
+    def __ge__(self, other):
+        return [obj >= other for obj in self]
+
+    def __add__(self, v):
+        if isinstance(v, AttrIter):
+            return AttrIter([x + y for x, y in zip(self, v)])
+        else:
+            return AttrIter([x + v for x in self])
+
+    def __sub__(self, v):
+        if isinstance(v, AttrIter):
+            return AttrIter([x - y for x, y in zip(self, v)])
+        else:
+            return AttrIter([x - v for x in self])
+
+    def __mul__(self, v):
[…]
