### 2021-04-30 f5bdae9
Update record & report methods
