+++ agentpy/objects.py
@@ -1,54 +1,30 @@
 """
 Agentpy Objects Module
-Content: Classes for environments and agents
+Content: Base classes for agents and environment
 """
 
-from .lists import AgentList, EnvList
+from .sequences import AgentList
 from .tools import AgentpyError, make_list
 
 
-class ApObj:
-    """ Agentpy base-class for objects of agent-based models."""
+class Object:
+    """ Base class for all objects of an agent-based models. """
 
     def __init__(self, model):
-        self._log = {}
-        self._model = model
         self._var_ignore = []
-        self._alive = True
-        self._id = model._new_id()  # Assign id to new object
-        self._model._obj_dict[self.id] = self  # Add object to object dict
+
+        self.id = model._new_id()  # Assign id to new object
+        self.type = type(self).__name__
+        self.log = {}
+
+        self.model = model
+        self.p = model.p
 
     def __repr__(self):
         return f"{self.type} (Obj {self.id})"
 
     def __getattr__(self, key):
         raise AttributeError(f"{self} has no attribute '{key}'.")
-
-    @property
-    def type(self):
-        return type(self).__name__
-
-    @property
-    def var_keys(self):
-        return [k for k in self.__dict__.keys()
-                if k[0] != '_'
-                and k not in self._var_ignore]
-
-    @property
-    def id(self):
-        return self._id
-
-    @property
-    def p(self):
-        return self._model._parameters
-
-    @property
-    def log(self):
-        return self._log
-
-    @property
-    def model(self):
-        return self._model
 
     def __getitem__(self, key):
         return getattr(self, key)
@@ -60,8 +36,16 @@ class ApObj:
         """Store current attributes to seperate them from custom variables"""
         self._var_ignore = [k for k in self.__dict__.keys() if k[0] != '_']
 
+    @property
+    def var_keys(self):
+        return [k for k in self.__dict__.keys()
+                if k[0] != '_'
+                and k not in self._var_ignore]
+
     def record(self, var_keys, value=None):
-        """ Records an objects variables.
+        """ Records an objects variables at the current time-step.
+        Recorded variables can be accessed via the object's `log` attribute
+        and will be saved to the model's output at the end of a simulation.
 
         Arguments:
             var_keys (str or list of str):
@@ -86,15 +70,28 @@ class ApObj:
                 a.record(a.var_keys)
         """
 
+        # Initial record call connects log to the models dict of logs
+        # and then overrides method with the actual recording function
+        if self.type not in self.model._logs:
+            self.model._logs[self.type] = {}
+        self.model._logs[self.type][self.id] = self.log
+        self.log['t'] = [self.model.t]  # Initiate time dimension
+        self.record = self._record  # noqa
+
+        for var_key in make_list(var_keys):
+
+            v = getattr(self, var_key) if value is None else value
+            self.log[var_key] = [v]
+
+    def _record(self, var_keys, value=None):
+
         for var_key in make_list(var_keys):
 
             # Create empty lists
-            if 't' not in self.log:
-                self.log['t'] = []
             if var_key not in self.log:
                 self.log[var_key] = [None] * len(self.log['t'])
 
-            if self.model.t not in self.log['t']:
+            if self.model.t != self.log['t'][-1]:
 
                 # Create empty slot for new documented time step
                 for v in self.log.values():
@@ -134,304 +131,31 @@ class ApObj:
             setattr(self, k, v)
 
 
-# Level 3 - Agent class
+class Environment(Object):
+    """ Base class for environments (Grid, Space, Network). """
 
-class Agent(ApObj):
-    """ Individual agent of an agent-based model.
+    pass
 
-    This class can be used as a parent class for custom agent types.
[…]
