+++ agentpy/model.py
@@ -5,228 +5,234 @@ Content: Main class for agent-based models
 
 import numpy as np
 import pandas as pd
-from datetime import datetime
+import random
 
-from .output import DataDict
-from .objects import ApEnv, Agent, Environment
+from os import sys
+from datetime import datetime
+from .version import __version__
+from .datadict import DataDict
+from .objects import Object, Environment
 from .network import Network
+from .sample import Range, Values
 from .grid import Grid
 from .space import Space
-from .tools import AttrDict, AgentpyError, make_list
-from .lists import ObjList, EnvList
+from .tools import AttrDict, AgentpyError, make_list, InfoStr
+from .sequences import AgentList
 
 
-class Model(ApEnv):
+class Model(Object):
     """
-    An agent-based model that can hold environments and agents.
-
-    This class can be used as a parent class for custom models.
-    Class attributes can be accessed like dictionary items.
-    To define the simulation procedure, you can override the methods
-    :func:`Model.setup`, :func:`Model.step`,
-    :func:`Model.update`, and :func:`Model.end`.
-    The perform the simulation, use :func:`Model.run`.
-
-    Attributes:
-        name (str): The models' name.
-        envs (EnvList): The models' environments.
-        agents (AgentList): The models' agents.
-        objects (ObjList): The models' agents and environments.
-        random (numpy.random.Generator): The models random number generator.
-        p (AttrDict): The models' parameters.
-        t (int): Current time-step of the model.
-        log (dict): The models' recorded variables.
-        measures (dict): The models' recorded measures.
-        var_keys (list): Names of the model's variables.
-        output (DataDict):
-            Output data that is generated at the end of a simulation.
+    Template of an agent-based model.
 
     Arguments:
-        parameters (dict, optional): Dictionary of model parameters.
-            Recommended types for parameters are int, float, str, list,
-            numpy.integer, numpy.floating, and numpy.ndarray.
-        run_id (int, optional): Number of current run (default None).
-        scenario (str, optional): Current scenario (default None).
+        parameters (dict, optional): Dictionary of the model's parameters.
         **kwargs: Will be forwarded to :func:`Model.setup`.
+
+    Attributes:
+        type (str): The model's class name.
+        info (str): Information about the model's current state.
+        p (AttrDict): The model's parameters.
+        t (int): Current time-step of the model.
+        id (int): The model's object id, which will always be zero.
+        random (random.Random): Random number generator.
+        nprandom (numpy.random.Generator): Numpy random number generator.
+        var_keys (list): Names of the model's custom variables.
+        running (bool): Indicates whether the model is currently running.
+        log (dict): The model's recorded variables.
+        reporters (dict): The model's documented reporters.
+        output (DataDict): Output data after a completed simulation.
+
+    Examples:
+
+        To define a custom model with a custom agent type::
+
+            class MyAgent(ap.Agent):
+
+                def setup(self):
+                    # Initialize an attribute with a parameter
+                    self.my_attribute = self.p.my_parameter
+
+                def agent_method(self):
+                    # Define custom actions here
+                    pass
+
+            class MyModel(ap.Model):
+
+                def setup(self):
+                    # Called at the start of the simulation
+                    self.agents = ap.AgentList(self, self.p.agents, MyAgent)
+
+                def step(self):
+                    # Called at every simulation step
+                    self.agents.agent_method()  # Call a method for every agent
+
+                def update(self):
+                    # Called after setup as well as after each step
+                    self.agents.record('my_attribute')  # Record variable
+
+                def end(self):
+                    # Called at the end of the simulation
[…]
