+++ agentpy/output.py
@@ -1,472 +1,0 @@
-"""
-Agentpy Output Module
-Content: DataDict class for output data
-"""
-
-import pandas as pd
-import os
-from os import listdir, makedirs
-from os.path import getmtime, join
-
-from .tools import AttrDict, make_list, AgentpyError
-import json
-import numpy as np
-
-
-class NpEncoder(json.JSONEncoder):
-    """ Adds support for numpy number formats to json. """
-    # By Jie Yang https://stackoverflow.com/a/57915246
-    def default(self, obj):
-        if isinstance(obj, np.integer):
-            return int(obj)
-        elif isinstance(obj, np.floating):
-            return float(obj)
-        elif isinstance(obj, np.ndarray):
-            return obj.tolist()
-        else:
-            return super(NpEncoder, self).default(obj)
-
-
-def _last_exp_id(name, path):
-    """ Identifies existing experiment data and return highest id. """
-
-    exp_id = 0
-    output_dirs = listdir(path)
-    exp_dirs = [s for s in output_dirs if name in s]
-    if exp_dirs:
-        ids = [int(s.split('_')[-1]) for s in exp_dirs]
-        exp_id = max(ids)
-    return exp_id
-
-
-class DataDict(AttrDict):
-    """ Dictionary for recorded simulation data.
-
-    Generated by :class:`Model`, :class:`Experiment`, or :func:`load`.
-    Dictionary items can be defined and accessed like attributes.
-    Attributes can differ from the standard ones listed below.
-
-    Attributes:
-        log (dict): Meta-data of the simulation
-            (e.g. name, time-stamps, settings, etc.).
-        parameters (dict, pandas.DataFrame, or DataDict):
-            Parameters that have been used for the simulation.
-        variables (pandas.DataFrame or DataDict)):
-            Dynamic variables, seperated per object type,
-            which can be recorded once per time-step with :func:`record`.
-        measures (pandas.DataFrame): Evaluation measures,
-            which can be recorded once per run with :func:`measure`.
-    """
-
-    def __repr__(self, indent=False):
-        rep = ""
-        if not indent:
-            rep += "DataDict {"
-        i = '    ' if indent else ''
-        for k, v in self.items():
-            rep += f"\n{i}'{k}': "
-            if isinstance(v, (int, float, np.integer, np.floating)):
-                rep += f"{v} {type(v)}"
-            elif isinstance(v, str):
-                x0 = f"(length {len(v)})"
-                x = f"...' {x0}" if len(v) > 20 else "'"
-                rep += f"'{v[:30]}{x} {type(v)}"
-            elif isinstance(v, pd.DataFrame):
-                lv = len(list(v.columns))
-                rv = len(list(v.index))
-                rep += f"DataFrame with {lv} " \
-                       f"variable{'s' if lv != 1 else ''} " \
-                       f"and {rv} row{'s' if rv != 1 else ''}"
-            elif isinstance(v, DataDict):
-                rep += f"{v.__repr__(indent=True)}"
-            elif isinstance(v, dict):
-                lv = len(list(v.keys()))
-                rep += f"Dictionary with {lv} key{'s' if lv != 1 else ''}"
-            elif isinstance(v, list):
-                lv = len(v)
-                rep += f"List with {lv} entr{'ies' if lv != 1 else 'y'}"
-            else:
-                rep += f"Object of type {type(v)}"
-        if not indent:
-            rep += "\n}"
-        return rep
-
-    def _short_repr(self):
-        len_ = len(self.keys())
-        return f"DataDict {{{len_} entr{'y' if len_ == 1 else 'ies'}}}"
-
-    def __eq__(self, other):
-        """ Check equivalence of two DataDicts."""
-        if not isinstance(other, DataDict):
-            return False
-        for key, item in self.items():
[…]
