+++ tests/test_output.py
@@ -1,321 +1,0 @@
-import pytest
-import agentpy as ap
-import numpy as np
-import pandas as pd
-import shutil
-import os
-
-from agentpy.tools import AgentpyError
-
-
-def test_combine_vars():
-
-    model = ap.Model()
-    model.record('test', 1)
-    results = model.run(1, display=False)
-    assert results._combine_vars().shape == (1, 1)
-
-    model = ap.Model()
-    agents = model.add_agents()
-    agents.record('test', 1)
-    results = model.run(1, display=False)
-    assert results._combine_vars().shape == (1, 1)
-
-    model = ap.Model()
-    agents = model.add_agents()
-    model.record('test', 1)
-    agents.record('test', 2)
-    results = model.run(1, display=False)
-    assert results._combine_vars().shape == (2, 1)
-
-    model = ap.Model()
-    agents = model.add_agents()
-    model.record('test', 1)
-    agents.record('test', 2)
-    results = model.run(1, display=False)
-    assert results._combine_vars(obj_types="Model").shape == (1, 1)
-
-    model = ap.Model()
-    agents = model.add_agents()
-    model.record('test', 1)
-    agents.record('test', 2)
-    results = model.run(1, display=False)
-    assert results._combine_vars(obj_types="Doesn't exist") is None
-
-    model = ap.Model()
-    results = model.run(1, display=False)
-    assert results._combine_vars() is None
-    assert results._combine_pars() is None
-
-    model = ap.Model({'test': 1})
-    results = model.run(1, display=False)
-    assert results._combine_pars(fixed=False) is None
-
-    results.variables = 1
-    with pytest.raises(TypeError):
-        assert results._combine_vars()
-    results.parameters = 1
-    with pytest.raises(TypeError):
-        assert results._combine_pars()
-
-
-def test_arrange_scenarios():
-
-    class MyModel(ap.Model):
-        def step(self):
-            self.record('test', 1)
-            self.measure('test', 1)
-            self.stop()
-
-    exp = ap.Experiment(MyModel, scenarios=['sc1', 'sc2'], record=True)
-    results = exp.run()
-
-    assert results.arrange() is None
-    assert results.arrange(variables='all', scenarios='sc1').shape == (1, 4)
-    assert results.arrange(measures='all', scenarios='sc1').shape == (1, 3)
-
-
-repr = """DataDict {
-'parameters': 
-    'fixed': Dictionary with 1 key
-    'varied': DataFrame with 1 variable and 10 rows
-'log': Dictionary with 5 keys
-'measures': DataFrame with 1 variable and 10 rows
-'variables': 
-    'Agent': DataFrame with 1 variable and 10 rows
-    'MyModel': DataFrame with 1 variable and 10 rows
-}"""
-
-
-class MyModel(ap.Model):
-    def step(self):
-        self.measure('x', self.p.x)
-        self.add_agents()
-        self.agents.record('id')
-        self.record('id')
-        self.stop()
-
-
-def test_repr():
-    param_ranges = {'x': (0., 1.), 'y': 1}
-    sample = ap.sample(param_ranges, n=10)
-    results = ap.Experiment(MyModel, sample, record=True).run()
-    assert results.__repr__() == repr
-
-
-class AgentType1(ap.Agent):
-    def setup(self):
-        self.x = 'x1'
-
-    def action(self):
-        self.record('x')
-
-
-class AgentType2(AgentType1):
-    def setup(self):
-        self.x = 'x2'
-        self.y = 'y2'
-
-    def action(self):
-        self.record(['x', 'y'])
-
-
[…]
