+++ docs/reference_models.rst
@@ -1,9 +1,0 @@
-.. currentmodule:: agentpy
-
-==================
-Agent-based models
-==================
-
-.. autoclass:: Model
-    :members:
-    :inherited-members:
