+++ docs/reference_experiments.rst
@@ -1,8 +1,0 @@
-.. currentmodule:: agentpy
-
-===========
-Experiments
-===========
-
-.. autoclass:: Experiment
-    :members:

[…]
