### 2021-04-30 4f59837
New class Sample
