+++ agentpy/sample.py
@@ -1,0 +1,198 @@
+"""
+Agentpy Sampling Module
+Content: Sampling functions
+"""
+
+# TODO Latin Hypercube
+# TODO Random distribution samples
+# TODO Store meta-info for later analysis
+
+import itertools
+import numpy as np
+
+from SALib.sample import saltelli
+from .tools import param_tuples_to_salib, InfoStr
+
+
+class Range:
+    """ A range of parameter values
+    that can be used to create a :class:`Sample`.
+
+    Arguments:
+        vmin (int or float, optional):
+            Minimum value for this parameter (default 0).
+        vmax (int or float, optional):
+            Maximum value for this parameter (default 1).
+        vdef (int or float, optional):
+            Default value. If none is passed, `(vmin+vmax)/2` is used.
+        ints (bool, optional):
+            Round and convert sampled values to integers (default False).
+    """
+
+    def __init__(self, vmin=0, vmax=1, vdef=None, ints=False):
+        self.vmin = vmin
+        self.vmax = vmax
+        self.vdef = vdef if vdef else (vmin+vmax)/2
+        self.ints = ints
+
+    def __repr__(self):
+        return f"Parameter range from {self.vmin} to {self.vmax}"
+
+
+class Values:
+    """ A pre-defined set of discrete parameter values
+    that can be used to create a :class:`Sample`.
+
+    Arguments:
+        *args:
+            Possible values for this parameter.
+        vdef:
+            Default value. If none is passed, the first passed value is used.
+    """
+
+    def __init__(self, *args, vdef=None):
+        self.values = args
+        self.vdef = vdef if vdef else args[0]
+
+    def __len__(self):
+        return len(self.values)
+
+    def __repr__(self):
+        return f"Set of {len(self.values)} parameter values"
+
+
+class Sample:
+    """ A sequence of parameter combinations
+    that can be used for :class:`Experiment`.
+
+    Arguments:
+
+        parameters (dict):
+            Dictionary of parameter keys and values.
+            Entries of type :class:`Range` and :class:`Values`
+            will be sampled based on chosen `method` and `n`.
+            Other types wil be interpreted as constants.
+
+        n (int, optional):
+            Sampling factor used by chosen `method` (default None).
+
+        method (str, optional):
+            Method to use to create parameter combinations
+            from entries of type :class:`Range`. Options are:
+
+            - ``linspace`` (default):
+              Arange `n` evenly spaced values for each :class:`Range`
+              and combine them with given :class:`Values` and constants.
+              Additional keyword arguments:
+
+                - ``product`` (bool, optional):
+                  Return all possible combinations (default True).
+                  If False, value sets are 'zipped' so that the i-th
+                  parameter combination contains the i-th entry of each
+                  value set. Requires all value sets to have the same length.
+
+            - ``saltelli``:
+              Apply Saltelli's sampling scheme,
+              using :func:`SALib.sample.saltelli.sample` with `N=n`.
+              This enables the analysis of Sobol Sensitivity Indices
+              with :func:`DataDict.calc_sobol` after the experiment.
+              Additional keyword arguments:
+
+                - ``calc_second_order`` (bool, optional):
+                  Whether to calculate second-order indices (default False).
+
+        **kwargs: Additional keyword arguments for chosen `method`.
+
+    """
+    def __init__(self, parameters, n=None, method='linspace', **kwargs):
+
+        self._type = method
+        self._sample = getattr(self, f"_{method}")(parameters, n, **kwargs)
+
+    def __repr__(self):
+        return f"Sample of {len(self)} parameter combinations"
+
+    def __iter__(self):
+        return iter(self._sample)
+
+    def __len__(self):
[…]
