+++ agentpy/space.py
@@ -3,220 +3,180 @@ Agentpy Space Module
 Content: Class for continuous spatial environments
 """
 
-# There is much room for optimization in this module.
-# Contributions are welcome! :)
-
 # TODO Add option of space without shape (infinite)
-# TODO Improve datatype consistency (list, np.array, etc.)
-# TODO Method for distance between agents
-# TODO Create method items() consistent with Grid
+# TODO Custom iterator for neighbors() & select() for performance
 
 import itertools
 import numpy as np
 import random as rd
 import collections.abc as abc
 from scipy import spatial
-from .objects import ApEnv, Agent
+from .objects import Spatial
 from .tools import make_list, make_matrix
-from .lists import AgentList
+from .sequences import AgentList, AgentIter
 
 
-class Space(ApEnv):
+class Space(Spatial):
     """ Environment that contains agents with a continuous spatial topology.
     To add new space environments to a model, use :func:`Model.add_space`.
-    For a discrete spatial topology, use :class:`Grid`.
+    For a discrete spatial topology, see :class:`Grid`.
 
     This class can be used as a parent class for custom space types.
     All agentpy model objects call the method :func:`setup` after creation,
     and can access class attributes like dictionary items.
-    See :class:`Environment` for general properties of all environments.
 
     Arguments:
         model (Model): The model instance.
         shape (tuple of float): Size of the space.
             The length of the tuple defines the number of dimensions,
             and the values in the tuple define the length of each dimension.
-        connect_borders (bool, optional):
+        torus (bool, optional):
             Whether to connect borders (default False).
-            If True, space will be toroidal, meaning that agents who, for
-            example, move over the right border, will appear on the left side.
+            If True, the space will be toroidal, meaning that agents who
+            move over a border will re-appear on the opposite side.
+            If False, they will remain at the edge of the border.
         **kwargs: Will be forwarded to :func:`Space.setup`.
 
     Attributes:
-        shape (tuple of float): Length of each spatial dimension.
-        dim (int): Number of dimensions.
+        agents (AgentIter):
+            Iterator over all agents in the space.
+        positions (dict of Agent):
+            Dictionary linking each agent instance to its position.
+        shape (tuple of float):
+            Length of each spatial dimension.
+        ndim (int):
+            Number of dimensions.
+        kdtree (scipy.spatial.cKDTree or None):
+            KDTree of agent positions for neighbor lookup.
+            Will be recalculated if agents have moved.
+            If there are no agents, tree is None.
     """
 
-    def __init__(self, model, shape, connect_borders=False, **kwargs):
+    def __init__(self, model, shape, torus=False, **kwargs):
 
         super().__init__(model)
 
-        self._topology = 'space'
-        self._connect_borders = connect_borders
+        self._torus = torus
         self._cKDTree = None
         self._sorted_agents = None
         self._sorted_agent_points = None
-        self._agent_dict = {}  # agent instance : agent position (np.array)
-        self._shape = tuple(shape)
+
+        self.positions = {}
+        self.shape = tuple(shape)
+        self.ndim = len(self.shape)
+
         self._set_var_ignore()
         self.setup(**kwargs)
 
     @property
-    def shape(self):
-        return self._shape
+    def agents(self):
+        return AgentIter(self.positions.keys())
 
     @property
-    def dim(self):
-        return len(self._shape)
-
-    @property
-    def KDTree(self):
-        """ KDTree of agent positions for neighbor lookup,
-        using :class:`scipy.spatial.KDTree`.
-        Tree is recalculated if agent's have moved or changed.
-        If there are no agents, returns None."""
+    def kdtree(self):
         # Create new KDTree if necessary
         if self._cKDTree is None and len(self.agents) > 0:
             self._sorted_agents = []
             self._sorted_agent_points = []
[…]
