+++ docs/reference_experiment.rst
@@ -1,0 +1,8 @@
+.. currentmodule:: agentpy
+
+===========
+Experiments
+===========
+
+.. autoclass:: Experiment
+    :members:
