+++ docs/reference_network.rst
@@ -1,0 +1,14 @@
+.. currentmodule:: agentpy
+
+=====================
+Network
+=====================
+
+.. autoclass:: Network
+    :members:
+    :inherited-members:
+
+.. autoclass:: AgentNode
+    :members:
+    :inherited-members:
+
