+++ docs/reference.rst
@@ -8,11 +8,12 @@ API Reference
 .. toctree::
    :maxdepth: 3
 
+   reference_model
    reference_agents
+   reference_sequences
    reference_environments
-   reference_models
-   reference_sampling
-   reference_experiments
+   reference_sample
+   reference_experiment
    reference_output
    reference_analysis
    reference_other
