+++ docs/changelog.rst
@@ -4,70 +4,90 @@
 Changelog
 =========
 
-Please note that API changes without backward compatibility
-are still possible until the release of v1.0.0.
+0.1.0.dev
+---------
 
-0.0.8.dev0
-----------
+This update contains some major revisions of most classes and methods in the
+package, including new features, better performance, and a more coherent syntax.
+Unfortunately, this causes some API changes without backward compatibility.
+There will hopefully be less major changes from now on,
+and backward compatibility will be ensured after first major release (v1.0.0).
 
-Grid and Network Locations
-..........................
+Object creation
+...............
 
-The environments :class:`Grid` and :class:`Network` have changed significantly,
-and now can now represent each grid cell or network node with a :class:`Location`.
-An agents' location can be accessed with the new method :func:`Agent.location`.
-Each location holds a number of agents, and can also have attributes itself.
+The methods :func:`add_agents`, :func:`add_env`, etc. have been depreciated.
+Instead, new objects can now be created directly or through sequences.
+For example::
 
-:obj:`Grid.grid` now returns a :class:`LocList` of the grids locations.
-Location attributes can be accessed like in an Agent List, e.g. to set an
-attribute x for all locations, one can call `my_grid.grid.x = 1`.
-To set an attribute x for all agents in those locations, one would call
-`my_grid.grid.agents.x = 1`.
+    class Model(ap.Model):
+        def setup(self):
+            self.special_agent = ap.Agent()  # Create a single agent
+            self.agents = ap.AgentList(self, 10)  # Create a sequence of 10 agents
+            self.grid = ap.Grid(self, (5, 5))  # Create a grid environment
 
+Data structures
+...............
 
-Improved lists
-..............
+The main advantage of the new way of object creation is increased control
+over data structures in which to hold objects. Some agents can be kept in
+different sequences than others, and different kinds of sequences can be used.
 
-Agent lists have a new method :func:`AgentList.call` that can be used to call
-a method for each agent in the list and has two optional arguments:
-`check_alive`, which prevents the call of agents that are deleted during the call loop;
-and `iter_kwargs`, which can be used to pass method arguments that are different for
-each agent.
+In addition to :class:`AgentList`, there is new sequence type
+:class:`AgentGroup` that provides increased performance
+for the lookup and deletion of agents.
+It also comes with a method `buffer` that allows for deletion of agents
+from the group during an active iteration over the group.
 
 The structure of :class:`AttrList` has been changed to an iterable over its
 source list. This improves performance and makes it possible to change
 agent attributes by setting new values to items in the attribute list (see
 :class:`AgentList` for an example). Otherwise, the class behaves as before.
 
-Agent-Environment interaction
-.............................
+Environments
+............
 
-The method :func:`Agent.neighbors` can now take multiple environments
-as an argument. If no environments are given, it will select all of the
-agents environments. For environments without a topology like :class:`Environment`
-and agents without any environments, an empty list will be returned.
-Agents further have a new attribute :obj:`alive` that indicates whether they
-have been removed from the model.
+The agent class has been split into two types :class:`Agent` and :class:`MultiAgent`,
+and can access their environment(s) through the attributes `pos` and `env`.
+The former should be used for models with zero or one environment, while the latter
+can be part of multiple environments.
 
-Other changes
+All three environment types have been revised.
+:class:`Grid` is now a structured array
+that can hold field attributes per position in addition to agents,
+and can be customized with the arguments `torus`, `track_free`, and `check_border`.
+:class:`Network` can now hold multiple agents per node.
+
+Random number generators
+........................
+
[…]
