+++ tests/test_objects.py
@@ -1,53 +1,35 @@
 import pytest
 import agentpy as ap
 
-repr = """Agent-based model {
-'type': Model
-'agents': AgentList [1 agent]
-'envs': EnvList [1 environment]
-'p': AttrDict {0 entries}
-'t': 0
-'run_id': None
-'scenario': None
-'output': DataDict {1 entry}
-}"""
-
 
 def test_basics():
     model = ap.Model()
-    env = model.add_env()
-    env.add_agents()
-    agent = model.agents[0]
+    agent = ap.Agent(model)
     agent.x = 1
     agent['y'] = 2
     assert agent['x'] == 1
     assert agent.y == 2
-    assert env.__repr__() == "Environment (Obj 1)"
-    assert agent.__repr__() == "Agent (Obj 2)"
-    assert env.topology is None
+    assert agent.__repr__() == "Agent (Obj 1)"
     assert model.type == 'Model'
-    assert model.__repr__() == repr
+    assert model.__repr__() == "Model"
+    assert isinstance(model.info, ap.tools.InfoStr)
     with pytest.raises(AttributeError):
         assert agent.z
-    assert agent is model.get_obj(2)
-    with pytest.raises(ValueError):
-        assert model.get_obj(3)
 
 
 def test_record():
     """ Record a dynamic variable """
 
     model = ap.Model()
-    model.add_agents(3)
     model.var1 = 1
     model.var2 = 2
     model.record(['var1', 'var2'])
     model.record('var3', 3)
 
-    assert len(list(model._log.keys())) == 3 + 1  # One for time
-    assert model._log['var1'] == [1]
-    assert model._log['var2'] == [2]
-    assert model._log['var3'] == [3]
+    assert len(list(model.log.keys())) == 3 + 1  # One for time
+    assert model.log['var1'] == [1]
+    assert model.log['var2'] == [2]
+    assert model.log['var3'] == [3]
 
 
 def test_record_all():
@@ -58,6 +40,6 @@ def test_record_all():
     model.var2 = 2
     model.record(model.var_keys)
 
-    assert len(list(model._log.keys())) == 3
-    assert model._log['var1'] == [1]
-    assert model._log['var2'] == [2]
+    assert len(list(model.log.keys())) == 3
+    assert model.log['var1'] == [1]
+    assert model.log['var2'] == [2]

[…]
