+++ docs/reference_sequences.rst
@@ -8,11 +8,12 @@ This module offers various data structures to create and manage groups
 of both agents and environments. Which structure best to use
 depends on the specific requirements of each model.
 
-- :class:`AgentList` provides an enhanced list with
+- :class:`AgentList` is a list of agentpy objects with
   methods to select and manipulate its entries.
-- :class:`AgentGroup` offers additional features
-  and better performance for removing and looking up objects.
-- :class:`AgentSet` is a simple unordered collection of agents.
+- :class:`AgentGroup` is an ordered collection of agentpy objects,
+  optimized for removing and looking up objects.
+- :class:`AgentSet` is an unordered collection of agents
+  that can access agent attributes.
 - :class:`AgentIter` and :class:`AgentGroupIter` are a list-like iterators
   over a selection of agentpy objects.
 - :class:`AttrIter` is a list-like iterator over the attributes of
