### 2021-04-30 dff72bc
Same methods for group and list
