+++ agentpy/sequences.py
@@ -8,11 +8,11 @@ Content: Lists for objects, environments, and agents
 import itertools
 import agentpy as ap
 import numpy as np
-from .tools import AgentpyError
-from collections.abc import Sequence as PySequence
+from .tools import AgentpyError, ListDict
+from collections.abc import Sequence
 
 
-class Sequence:
+class AgentSequence:
     """ Base class for agenpty sequences. """
 
     def __repr__(self):
@@ -20,10 +20,10 @@ class Sequence:
         return f"{type(self).__name__} ({len(self)} object{s})"
 
     def __getattr__(self, name):
-        # TODO This breaks numpy conversion because of __array_struct__ lookup
         """ Return callable list of attributes """
-        if name[0] == '_':  # Allows for numpy conversion
-            super().__getattr__(name)  #raise AttributeError('test')
+        if name[0] == '_':  # Private variables are looked up normally
+            # Gives numpy conversion correct error for __array_struct__ lookup
+            super().__getattr__(name)
         else:
             return AttrIter(self, attr=name)
 
@@ -40,7 +40,7 @@ class Sequence:
 
 # Attribute List ------------------------------------------------------------ #
 
-class AttrIter(Sequence, PySequence):
+class AttrIter(AgentSequence, Sequence):
     """ Iterator over an attribute of objects in a sequence.
     Length, items access, and representation work like with a normal list.
     Calls are forwarded to each entry and return a list of return values.
@@ -142,7 +142,28 @@ class AttrIter(Sequence, PySequence):
 
 # Object Containers --------------------------------------------------------- #
 
-class AgentList(Sequence, list):
+def _random(gen, obj_list, n=1, replace=False):
+    """ Creates a random sample of agents.
+
+    Arguments:
+        n (int, optional): Number of agents (default 1).
+        replace (bool, optional):
+            Select with replacement (default False).
+            If True, the same agent can be selected more than once.
+
+    Returns:
+        AgentIter: The selected agents.
+    """
+    if n == 1:
+        selection = [gen.choice(obj_list)]
+    elif replace is False:
+        selection = gen.sample(obj_list, k=n)
+    else:
+        selection = gen.choices(obj_list, k=n)
+    return AgentIter(selection)
+
+
+class AgentList(AgentSequence, list):
     """ List of agentpy objects.
     Attribute calls and assignments are applied to all agents
     and return an :class:`AttrIter` with the attributes of each agent.
@@ -237,35 +258,19 @@ class AgentList(Sequence, list):
         """
         return AgentList(self.model, [a for a, s in zip(self, selection) if s])
 
-    # TODO offer both choices
-    def random(self, n=1, replace=False, weights=None, shuffle=True):
-        """ Creates a random sample of agents,
-        using :func:`numpy.random.Generator.choice`.
-        Argument descriptions are adapted from :obj:`numpy.random`.
-        Returns a new :class:`AgentList` with the selected agents.
+    def random(self, n=1, replace=False):
+        """ Creates a random sample of agents.
 
         Arguments:
             n (int, optional): Number of agents (default 1).
             replace (bool, optional):
-                Whether the sample is with or without replacement.
-                Default is False, meaning that every agent can
-                only be selected once.
-            weights (1-D array_like, optional):
-                The probabilities associated with each agent.
-                If not given the sample assumes a uniform distribution
-                over all agents.
-            shuffle (bool, optional):
-                Whether the sample is shuffled
-                when sampling without replacement.
-                Default is True, False provides a speedup.
-        """
+                Select with replacement (default False).
+                If True, the same agent can be selected more than once.
 
[…]
