+++ docs/overview.rst
@@ -16,16 +16,16 @@ Creating models
 
 The basic framework for agent-based models consists of three levels:
 
-1. :class:`Model`, which contains agents, environments, parameters, & procedures
-2. :class:`Environment`, :class:`Grid`, :class:`Space`, and :class:`Network`, which contain agents
-3. :class:`Agent`, the basic building blocks of the model
+1. The :class:`Model` itself, which contains agents, environments, parameters, & procedures
+2. The environment types :class:`Grid`, :class:`Space`, and :class:`Network`, which can contain agents
+3. The agent types :class:`Agent` and :class:`MultiAgent`, the basic building blocks of the model
 
 All of these classes are designed to be customized through the creation of
 `sub-classes <https://docs.python.org/3/tutorial/classes.html?highlight=inheritance#inheritance>`_
 with their own variables and methods.
-A custom agent type could be defined as follows::
+A custom agent type can be defined as follows::
 
-    class MyAgentType(ap.Agent):
+    class MyAgent(ap.Agent):
 
         def setup(self):
             # Initialize an attribute with a parameter
@@ -60,7 +60,7 @@ here is how a basic model could look like::
 
         def setup(self):
             """ Called at the start of the simulation """
-            self.add_agents(self.p.agents, MyAgentType)  # Add new agents
+            self.agents = ap.AgentList(self, self.p.agents, MyAgent)  # New agents
 
         def step(self):
             """ Called at every simulation step """
@@ -85,10 +85,9 @@ Using agents
 ############
 
 Agentpy comes with various tools to create, manipulate, and delete agents.
-The method :func:`Model.add_agents` can be used to initialize new agents.
-A list of all agents in a model can be accessed through :attr:`Model.agents`.
-Lists of agents are returned as an :class:`AgentList`,
-which provides special features to access and manipulate the whole group of agents.
+The classes :class:`AgentList` and :class:`AgentGroup`
+can be used to initialize and contain agents.
+These sequences provide special features to access and manipulate the whole group of agents.
 
 For example, when the model defined above calls :func:`self.agents.agent_method`,
 it will call the method :func:`MyAgentType.agent_method` for every agent in the model.
@@ -96,9 +95,9 @@ Similar commands can be used to set and access variables, or select subsets
 of agents with boolean operators.
 The following command, for example, would select all agents with an id above one::
 
-    self.agents.select(self.agents.id > 1)
+    agents.select(agents.id > 1)
 
-Further examples can be found in the :class:`AgentList` reference
+Further examples can be found in the :doc:`reference_sequences` reference
 or the :doc:`agentpy_virus_spread` model.
 
 .. _overview_environments:
@@ -109,15 +108,8 @@ Using environments
 Environments can contain agents just like the main model,
 and are useful if one wants to regard particular topologies for interaction
 or multiple environments that can hold seperate populations of agents.
-Agents can be moved between environments with the methods
-:func:`Agent.enter` and :func:`Agent.exit`.
-
-New environments can be created with :func:`Model.add_env`.
-Similar to agents, the attribute :attr:`envs` returns an :class:`EnvList`
-with special features to deal with groups of environments.
 There are three different types of environments:
 
-- :class:`Environment`, which simply contain agents without any topology.
 - :class:`Network`, in which agents can be connected via a networkx graph.
 - :class:`Grid`, in which agents occupy a position on a discrete x-dimensional space.
 - :class:`Space`, in which agents occupy a position on a continuous x-dimensional space.
@@ -137,10 +129,11 @@ which can be stored for each object (agent, environment, or model) and time-step
[…]
