+++ agentpy/datadict.py
@@ -7,7 +7,7 @@ import pandas as pd
 import os
 from os import listdir, makedirs
 from os.path import getmtime, join
-
+from SALib.analyze import sobol
 from .tools import AttrDict, make_list, AgentpyError
 import json
 import numpy as np
@@ -111,54 +111,79 @@ class DataDict(AttrDict):
     def __ne__(self, other):
         return not self.__eq__(other)
 
-    # Data manipulation ----------------------------------------------------- #
-
-    def merge_iterations(self):
-        pass
-
     # Data analysis --------------------------------------------------------- #
 
+    def calc_average_reporters(self, method='mean'):
+        """ Calculates average reporters over multiple iterations
+        and return self with a new entry 'average_reporters'.
+
+        Arguments:
+            method (str, optional):
+                Method to use for average.
+                Options are 'mean' (default), 'median', and 'mode'.
+        """
+        df = self.reporters
+        if method == 'mean':
+            ndf = df.groupby('sample_id').mean()
+        elif method == 'median':
+            ndf = df.groupby('sample_id').median()
+        elif method == 'mode':
+            ndf = df.groupby('sample_id').mode()
+        else:
+            raise AgentpyError("Method must be 'mean', 'median', or 'mode'.")
+        self['average_reporters'] = ndf
+        return self
+
+    @staticmethod
     def _sobol_set_df_index(df, p_keys, reporter):
         df['parameter'] = p_keys
         df['reporter'] = reporter
         df.set_index(['reporter', 'parameter'], inplace=True)
 
-    def calc_sobol(self, param_ranges, reporters=None,
-                         calc_second_order=False, **kwargs):
-        """ Calculates Sobol Sensitivity Indices and adds them to the output,
-        using :func:`SALib.analyze.sobol.analyze`. To work correctly,
-        data must be from an experiment with a parameter sample that was
-        generated by :func:`sample_saltelli`.
+    def calc_sobol(self, reporters=None, **kwargs):
+        """ Calculates Sobol Sensitivity Indices
+        using :func:`SALib.analyze.sobol.analyze`.
+        Data must be from an :class:`Experiment` with a :class:`Sample`
+        that was generated with the method 'saltelli'.
+        If the experiment had more than one iteration,
+        average reporters must first be calculated with
+        :func:`DataDict.calc_average_reporters`.
 
         Arguments:
-            output (DataDict): The output of an experiment that was set to only
-                one iteration (default) and used a parameter sample that was
-                generated with :func:`sample_saltelli`.
-            param_ranges (dict): The same dictionary that was used for the
-                generation of the parameter sample with :func:`sample_saltelli`.
             reporters (str or list of str, optional): The reporters that should
                 be used for the analysis. If none are passed, all are used.
-            calc_second_order (bool, optional): Whether to calculate second order
-                indices (default False). Value must be the same as the one used
-                to generate the parameter sample with :func:`sample_saltelli`.
             **kwargs: Will be forwarded to :func:`SALib.analyze.sobol.analyze`.
+
+        Returns:
+            DataDict: The DataDict itself with an added category 'sensitivity'.
         """
 
-        # STEP 1 - Convert param_ranges to SALib Format
-        param_ranges_tuples = {k: v for k, v in param_ranges.items()
-                               if isinstance(v, tuple)}
-        param_ranges_salib = param_tuples_to_salib(param_ranges_tuples)
+        if not self.parameters.sample_log['type'] == 'saltelli':
[…]
