+++ docs/reference_output.rst
@@ -5,6 +5,24 @@ Output data
 ===========
 
 .. autoclass:: DataDict
-    :members:
+
+Data arrangement
+################
+
+.. automethod:: DataDict.arrange
+.. automethod:: DataDict.arrange_reporters
+.. automethod:: DataDict.arrange_variables
+
+Analysis methods
+################
+
+.. automethod:: DataDict.calc_average_reporters
+
+.. automethod:: DataDict.calc_sobol
+
+Saving and loading
+##################
+
+.. automethod:: DataDict.save
 
 .. autofunction:: load
