### 2021-05-05 JoelForamitti/agentpy#15 @martibosch
[JOSS Review] Scenario comparison example
The docstring of the `scenarios` argument in `Experiment` does not give enough information, I have to go to the code to understand how it works. Additionally, I would highly recommend including the "scenario comparison" functionality in one of the example notebooks.
