### 2021-05-05 JoelForamitti/agentpy#16 @martibosch
[JOSS review] Manuscript corrections
## Statement of need

AFAIK, this section *must* be explicit in the manuscript. I suppose that such a section would include the sentence of lines 13-14. I would further suggest to make it a "self-contained" sentence (that does not require reading any other part of the manuscript to fully understand it), hence I would add:

    The aim of agentpy is to provide an intuitive syntax for the creation of **agent-based** models together with
    advanced tools for scientific applications.

## Figure and code snippets

* Figure 1: it may be better to use a code snippet rather than a figure to display code, e.g., users might want to copy-paste them
* lines 27-30: it is better that the text refers to some code snippet in the manuscript (e.g., Figure 1). For instance: from the current manuscript, one may ask: what is `agents`? I had to open a notebook, instantiate some example `agentpy.Model`, inspect the class via `dir` and it is only then that I found the `agents` attribute

## Clarify some concepts and definitions

* line 30: what is the "record" action? I needed to go to the docstring of the `record` method to fully understand what it means. It may be a good idea to add a brief sentence to clarify it *within the manuscript* so that the reader can get a clearer picture there.
* line 45: "direct integration of analysis tools within the same framework" as somebody familiar with scientific Python, I suppose that you mean NumPy, pandas, NetworkX and the like, but it may be better to make it more explicit.

## Other

* line 29: IMHO "attribute" seems more Pythonic than "variable"
* links: it might be better to put URLs explicitly (e.g., although unlikely in JOSS, people might print the paper). Line 18 is fine since we can easily find the documentation without clicking the link, but not so much for line 46, where I believe that the word "here" should be replaced by the actual URL.
