### 2021-05-05 #3065 @martibosch
Hello all! I am sorry for my delay, these were (and still are) some crazy days with the end of my thesis and some proposal deadlines to find further funding opportunities.

I have finished my review, I find the package to be very neat and easy (enjoyable) to use, which is great. The examples are clear and properly cover all the features except the scenario comparison, which is why I submitted [a dedicated issue](https://github.com/JoelForamitti/agentpy/issues/15). Besides this, I only have some minor corrections in the manuscript which are addressed in [another issue](https://github.com/JoelForamitti/agentpy/issues/16).

Overall, I find that agentpy is a great piece of work, my remarks are minor and should be quite easy to address. Looking forward to use it for my research!

Best,
Martí
