+++ docs/reference_analysis.rst
@@ -1,15 +1,0 @@
-.. currentmodule:: agentpy
-
-========
-Analysis
-========
-
-Animations
-----------
-
-.. autofunction:: animate
-
-Plots
------
-
-.. autofunction:: gridplot
