+++ docs/reference_output.rst
@@ -1,28 +1,0 @@
-.. currentmodule:: agentpy
-
-===========
-Output data
-===========
-
-.. autoclass:: DataDict
-
-Data arrangement
-################
-
-.. automethod:: DataDict.arrange
-.. automethod:: DataDict.arrange_reporters
-.. automethod:: DataDict.arrange_variables
-
-Analysis methods
-################
-
-.. automethod:: DataDict.calc_average_reporters
-
-.. automethod:: DataDict.calc_sobol
-
-Saving and loading
-##################
-
-.. automethod:: DataDict.save
-
-.. autofunction:: load
