### 2021-05-10 e60ffaf
Revision DataDict
log to info, load as class method, improved calc_sobol, removed calc_averag
