+++ agentpy/datadict.py
@@ -40,21 +40,21 @@ def _last_exp_id(name, path):
 
 
 class DataDict(AttrDict):
-    """ Dictionary for recorded simulation data.
-
-    Generated by :class:`Model`, :class:`Experiment`, or :func:`load`.
-    Dictionary items can be defined and accessed like attributes.
+    """ Nested dictionary for output data of simulations.
+    Items can be accessed like attributes.
     Attributes can differ from the standard ones listed below.
 
     Attributes:
-        log (dict):
-            Meta-data of the simulation.
+        info (dict):
+            Metadata of the simulation.
         parameters (dict or DataDict):
-            Parameters that have been used for the simulation.
-        variables (DataDict)):
+            Simulation parameters.
+        variables (DataDict):
             Recorded variables, seperated per object type.
         reporters (pandas.DataFrame):
             Reported outcomes of the simulation.
+        sensitivity (DataDict):
+            Sensitivity data, if calculated.
     """
 
     def __repr__(self, indent=False):
@@ -113,27 +113,6 @@ class DataDict(AttrDict):
 
     # Data analysis --------------------------------------------------------- #
 
-    def calc_average_reporters(self, method='mean'):
-        """ Calculates average reporters over multiple iterations
-        and return self with a new entry 'average_reporters'.
-
-        Arguments:
-            method (str, optional):
-                Method to use for average.
-                Options are 'mean' (default), 'median', and 'mode'.
-        """
-        df = self.reporters
-        if method == 'mean':
-            ndf = df.groupby('sample_id').mean()
-        elif method == 'median':
-            ndf = df.groupby('sample_id').median()
-        elif method == 'mode':
-            ndf = df.groupby('sample_id').mode()
-        else:
-            raise AgentpyError("Method must be 'mean', 'median', or 'mode'.")
-        self['average_reporters'] = ndf
-        return self
-
     @staticmethod
     def _sobol_set_df_index(df, p_keys, reporter):
         df['parameter'] = p_keys
@@ -146,8 +125,7 @@ class DataDict(AttrDict):
         Data must be from an :class:`Experiment` with a :class:`Sample`
         that was generated with the method 'saltelli'.
         If the experiment had more than one iteration,
-        average reporters must first be calculated with
-        :func:`DataDict.calc_average_reporters`.
+        the mean value between iterations will be taken.
 
         Arguments:
             reporters (str or list of str, optional): The reporters that should
@@ -158,21 +136,16 @@ class DataDict(AttrDict):
             DataDict: The DataDict itself with an added category 'sensitivity'.
         """
 
-        if not self.parameters.sample_log['type'] == 'saltelli':
+        if not self.parameters.log['type'] == 'saltelli':
             raise AgentpyError("Sampling method must be 'saltelli'.")
-        if self.log['iterations'] == 1:
+        if self.info['iterations'] == 1:
             df = self.reporters
         else:
-            if 'average_reporters' in self:
-                df = self.average_reporters
-            else:
-                raise AgentpyError(
-                    "Average reporters must first be calculated with"
-                    "DataDict.calc_average_reporters.")
+            df = self.reporters.groupby('sample_id').mean()
 
         # STEP 1 - Load salib problem from parameter log
-        param_ranges_salib = self.parameters.sample_log['salib_problem']
-        calc_second_order = self.parameters.sample_log['calc_second_order']
+        param_ranges_salib = self.parameters.log['salib_problem']
+        calc_second_order = self.parameters.log['calc_second_order']
 
         # STEP 2 - Calculate Sobol Sensitivity Indices
         if reporters is None:
@@ -253,7 +226,7 @@ class DataDict(AttrDict):
                        if k in make_list(obj_types)}
 
         # Add 'obj_id' before 't' for model df
[…]
