### 2021-05-10 ecc50ca
Calc_second_order default True
