+++ agentpy/visualization.py
@@ -1,0 +1,128 @@
+"""
+Agentpy Visualization Module
+Content: Animations and Gridplot
+"""
+
+import matplotlib
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+
+from SALib.analyze import sobol
+
+from .tools import make_list, param_tuples_to_salib
+
+
+def animate(model, fig, axs, plot, steps=None, seed=None,
+            skip=0, fargs=(), **kwargs):
+    """ Returns an animation of the model simulation,
+    using :func:`matplotlib.animation.FuncAnimation`.
+
+    Arguments:
+        model (Model): The model instance.
+        fig (matplotlib.figure.Figure): Figure for the animation.
+        axs (matplotlib.axes.Axes or list): Axis or list of axis of the figure.
+        plot (function): Function that takes the arguments `model, axs, *fargs`
+            and creates the desired plots on each axis at each time-step.
+        steps(int, optional):
+            Maximum number of steps for the simulation to run.
+            If none is given, the parameter 'Model.p.steps' will be used.
+            If there is no such parameter, 'steps' will be infinite.
+        seed (int, optional):
+            Seed for the models random number generators.
+            If none is given, the parameter 'Model.p.seed' will be used.
+            If there is no such parameter, a random seed will be used.
+        skip (int, optional):
+            Steps to skip before the animation starts (default 0).
+        fargs (tuple, optional): Forwarded fo the `plot` function.
+        **kwargs: Forwarded to :func:`matplotlib.animation.FuncAnimation`.
+
+    Examples:
+        An animation can be generated as follows::
+
+            def my_plot(model, ax):
+                pass  # Call pyplot functions here
+            
+            fig, ax = plt.subplots() 
+            my_model = MyModel(parameters)
+            animation = ap.animate(my_model, fig, ax, my_plot)
+
+        One way to display the resulting animation object in Jupyter::
+
+            from IPython.display import HTML
+            HTML(animation.to_jshtml())
+    """
+
+    model.run_setup(steps, seed)
+    model.create_output()
+    pre_steps = 0
+
+    for _ in range(skip):
+        model.run_step()
+
+    def frames():
+        nonlocal model, pre_steps
+        if model.running is True:
+            while model.running:
+                if pre_steps < 2:  # Frames iterates twice before starting plot
+                    pre_steps += 1
+                else:
+                    model.run_step()
+                    model.create_output()
+                yield model.t
+        else:  # Yield current if model stops before the animation starts
+            yield model.t
+
+    def update(t, m, axs, *fargs):  # noqa
+        nonlocal pre_steps
+        for ax in make_list(axs):
+            # Clear axes before each plot
+            ax.clear()
+        plot(m, axs, *fargs)  # Perform plot
+
+    ani = matplotlib.animation.FuncAnimation(
+        fig, update,
+        frames=frames,
+        fargs=(model, axs, *fargs),
+        save_count=model._steps,
+        **kwargs)  # noqa
+
+    plt.close()  # Don't display static plot
+    return ani
+
+
+def _apply_colors(grid, color_dict, convert):
+    if isinstance(grid[0], (list, np.ndarray)):
+        return [_apply_colors(subgrid, color_dict, convert)
+                for subgrid in grid]
+    else:
+        if color_dict is not None:
+            grid = [i if i is np.nan else color_dict[i] for i in grid]
+        if convert is True:
+            grid = [(0., 0., 0., 0.) if i is np.nan else
+                    matplotlib.colors.to_rgba(i) for i in grid]
+        return grid
+
+
[…]
