+++ docs/reference_visualization.rst
@@ -1,0 +1,9 @@
+.. currentmodule:: agentpy
+
+=============
+Visualization
+=============
+
+.. autofunction:: animate
+
+.. autofunction:: gridplot
