### 2021-05-13 ec5c1b5
Methods for MultiAgent
