### 2021-05-13 adba96a
Removed Environment and Spatial
