### 2021-05-13 6759ab8
New class IntRange
