### 2021-05-13 97767d0
New class IntRange
