+++ agentpy/__init__.py
@@ -10,12 +10,15 @@ Source: https://github.com/JoelForamitti/agentpy
 __all__ = [
     '__version__',
     'Model',
-    'Agent', 'AgentList', 'AgentGroup', 'AgentSet',
+    'Agent', 'MultiAgent',
+    'AgentList', 'AgentGroup', 'AgentSet',
     'AgentIter', 'AgentGroupIter', 'AttrIter',
-    'Grid', 'GridIter', 'Space', 'Network', 'AgentNode',
+    'Grid', 'GridIter',
+    'Space',
+    'Network', 'AgentNode',
     'Experiment',
     'DataDict',
-    'Sample', 'Values', 'Range',
+    'Sample', 'Values', 'Range', 'IntRange',
     'gridplot', 'animate',
     'AttrDict'
 ]
@@ -26,11 +29,11 @@ from .model import Model
 from .agent import Agent, MultiAgent
 from .sequences import AgentList, AgentGroup, AgentSet
 from .sequences import AgentIter, AgentGroupIter, AttrIter
-from .network import Network, AgentNode
 from .grid import Grid, GridIter
 from .space import Space
+from .network import Network, AgentNode
 from .experiment import Experiment
 from .datadict import DataDict
-from .sample import Sample, Values, Range
+from .sample import Sample, Values, Range, IntRange
 from .visualization import gridplot, animate
 from .tools import AttrDict
