### 2021-05-13 989e32a
Update torus and border behavior
