### 2021-05-13 f33ed94
Update simulation method names
