### 2021-05-13 81bb9b2
Object as Network parent class
