### 2021-05-13 310a2fd
Updated AgentIter repr
