### 2021-05-13 faac992
Update border behavior
