### 2021-05-13 f13d72a
Update simulation method names
