+++ agentpy/visualization.py
@@ -53,12 +53,12 @@ def animate(model, fig, axs, plot, steps=None, seed=None,
             HTML(animation.to_jshtml())
     """
 
-    model.run_setup(steps, seed)
+    model.sim_setup(steps, seed)
     model.create_output()
     pre_steps = 0
 
     for _ in range(skip):
-        model.run_step()
+        model.sim_step()
 
     def frames():
         nonlocal model, pre_steps
@@ -67,7 +67,7 @@ def animate(model, fig, axs, plot, steps=None, seed=None,
                 if pre_steps < 2:  # Frames iterates twice before starting plot
                     pre_steps += 1
                 else:
-                    model.run_step()
+                    model.sim_step()
                     model.create_output()
                 yield model.t
         else:  # Yield current if model stops before the animation starts
