+++ docs/reference_model.rst
@@ -7,7 +7,7 @@ Agent-based models
 The :class:`Model` contains all objects
 and defines the procedures of an agent-based simulation.
 It is meant as a template for custom model classes that
-can override the custom procedure methods (see below).
+override the `custom procedure methods`_.
 
 .. autoclass:: Model
 
@@ -16,6 +16,8 @@ Simulation tools
 
 .. automethod:: Model.run
 .. automethod:: Model.stop
+
+.. _custom procedure methods:
 
 Custom procedures
 #################
