+++ tests/test_objects.py
@@ -1,6 +1,6 @@
 import pytest
 import agentpy as ap
-
+from agentpy.tools import AgentpyError
 
 def test_basics():
     model = ap.Model()
@@ -15,6 +15,52 @@ def test_basics():
     assert isinstance(model.info, ap.tools.InfoStr)
     with pytest.raises(AttributeError):
         assert agent.z
+
+
+def test_single_env_agents():
+
+    model = ap.Model()
+    env = ap.Grid(model, (10, 10))
+    agent = ap.Agent(model)
+    env.add_agents([agent])
+
+    with pytest.raises(AgentpyError):
+        env.add_agents([agent])
+
+    assert env == agent.env
+    assert env.positions[agent] == agent.pos
+
+    env.remove_agents([agent])
+
+    assert agent.env is None
+    assert agent.pos is None
+
+    with pytest.raises(AgentpyError):
+        env.remove_agents([agent])
+
+
+def test_multi_env_agents():
+
+    model = ap.Model()
+    env1 = ap.Grid(model, (1, 1))
+    env2 = ap.Grid(model, (1, 1))
+    agent = ap.MultiAgent(model)
+    env1.add_agents([agent])
+    env2.add_agents([agent])
+
+    assert agent.env is agent.pos
+    assert agent.env == {env1: [0, 0], env2: [0, 0]}
+
+    env1.remove_agents([agent])
+
+    assert agent.env == {env2: [0, 0]}
+
+    env2.remove_agents([agent])
+
+    assert agent.env == {}
+
+    with pytest.raises(KeyError):
+        env2.remove_agents([agent])
 
 
 def test_record():
@@ -33,13 +79,14 @@ def test_record():
 
 
 def test_record_all():
-    """ Record all dynamic variables automatically """
+    """ Record all dynamic variables """
 
     model = ap.Model()
     model.var1 = 1
     model.var2 = 2
-    model.record(model.var_keys)
+    model.record(model.vars)
 
     assert len(list(model.log.keys())) == 3
     assert model.log['var1'] == [1]
     assert model.log['var2'] == [2]
+
