### 2021-05-13 JoelForamitti/agentpy#15 @JoelForamitti
Hi @martibosch, thank you for pointing this out.
Indeed, the scenarios argument has not made it into any of the demonstration models
as it has proven more straight-forward to represent scenarios through different parameter combinations.
It will be depreciated in the upcoming release of v0.1.0.
