### 2021-05-13 d9b8d63
Gridplot to numpy arrays
