### 2021-05-13 ec715bd
Doc revisions for JOSS
