+++ docs/overview.rst
@@ -4,25 +4,31 @@
 Overview
 ========
 
-This section aims to provide a rough overview over the main classes and
+This section provides an overview over the main classes and
 functions of agentpy and how they are meant to be used.
-For a more detailed description of each element, please refer to the :doc:`reference`.
+For a more detailed description of each element,
+please refer to the :doc:`guide` and :doc:`reference`.
 Throughout this documentation, agentpy is imported as follows::
 
     import agentpy as ap
 
+Structure
+#########
+
+The basic structure of the agentpy framework has four levels:
+
+1. The agent types :class:`Agent` and :class:`MultiAgent` are the basic building blocks of the model
+2. The environment types :class:`Grid`, :class:`Space`, and :class:`Network` contain agents
+3. A :class:`Model` contains agents, environments, parameters, and simulation procedures
+4. An :class:`Experiment` can run a model multiple times with different parameter combinations
+
+All of these classes are templates that can be customized through the creation of
+`sub-classes <https://docs.python.org/3/tutorial/classes.html?highlight=inheritance#inheritance>`_
+with their own variables and methods.
+
 Creating models
 ###############
 
-The basic framework for agent-based models consists of three levels:
-
-1. The :class:`Model` itself, which contains agents, environments, parameters, & procedures
-2. The environment types :class:`Grid`, :class:`Space`, and :class:`Network`, which can contain agents
-3. The agent types :class:`Agent` and :class:`MultiAgent`, the basic building blocks of the model
-
-All of these classes are designed to be customized through the creation of
-`sub-classes <https://docs.python.org/3/tutorial/classes.html?highlight=inheritance#inheritance>`_
-with their own variables and methods.
 A custom agent type can be defined as follows::
 
     class MyAgent(ap.Agent):
@@ -36,22 +42,17 @@ A custom agent type can be defined as follows::
             pass
 
 The method :func:`Agent.setup` is meant to be overwritten
-and will be called after an agents' creation.
-All variables of an agents should be initialized in this method.
+and will be called automatically after an agents' creation.
+All variables of an agents should be initialized within this method.
 Other methods can represent actions that the agent will be able to take during a simulation.
 
-We can further see that the agent comes with a built-in attribute :attr:`.p` that
-allows it to access the models' parameters.
-All model objects (i.e. agents, environments, and the model itself)
-are equipped with such properties to access different parts of the model:
+All model objects (including agents, environments, and the model itself)
+are equipped with the following default attributes:
 
-- :attr:`.model` returns the model instance
-- :attr:`.model.t` returns the model's time-step
-- :attr:`.id` returns a unique identifier number for each object
-- :attr:`.p` returns an :class:`AttrDict` of the models' parameters
-- :attr:`.envs` returns an :class:`EnvList` of the objects' environments
-- :attr:`.agents` (not for agents) returns an :class:`AgentList` of the objects' agents
-- :attr:`.log` returns a :class:`dict` of the objects' recorded variables
+- :attr:`model` the model instance
+- :attr:`id` a unique identifier number for each object
+- :attr:`p` the models' parameters
+- :attr:`log` the objects' recorded variables
 
 Using the new agent type defined above,
 here is how a basic model could look like::
@@ -59,100 +60,112 @@ here is how a basic model could look like::
     class MyModel(ap.Model):
 
         def setup(self):
-            """ Called at the start of the simulation """
-            self.agents = ap.AgentList(self, self.p.agents, MyAgent)  # New agents
+            """ Initiate a list of new agents. """
+            self.agents = ap.AgentList(self, self.p.agents, MyAgent)
 
         def step(self):
-            """ Called at every simulation step """
[…]
