+++ docs/guide.rst
@@ -1,14 +1,10 @@
 .. currentmodule:: agentpy
 
-==========
-User Guide
-==========
+===========
+User Guides
+===========
 
-Welcome to the agentpy user guide.
-This section contains various articles to help with specific problems and applications.
-Some of these articles are provided as interactive `Jupyter Notebooks <https://jupyter.org/>`_
-that can be downloaded and experimented with.
-
+This section contains interactive notebooks with common applications of the agentpy framework.
 If you are interested to add a new article to this guide, please visit :doc:`contributing`.
 If you are looking for examples of complete models, take a look at :doc:`model_library`.
 
