### 2021-05-13 79e6db7
Support 'nan' strings
