+++ tests/test_network.py
@@ -14,18 +14,22 @@ def test_add_agents():
     env = ap.Network(model, graph=graph)
     agents = ap.AgentList(model, 2)
     env.add_agents(agents, positions=env.nodes)
+    for agent in agents:
+        agent.pos = env.positions[agent]
     env.graph.add_edge(*agents.pos)
 
     # Test structure
     assert env.nodes == env.graph.nodes()
     assert list(env.graph.edges) == [tuple(agents.pos)]
-    assert list(agents[0].neighbors().id) == [3]
+    assert list(env.neighbors(agents[0]).id) == [3]
 
     # Add agents as new nodes
     model2 = ap.Model()
     agents2 = ap.AgentList(model2, 2)
     env2 = ap.Network(model2)
     env2.add_agents(agents2)
+    for agent in agents2:
+        agent.pos = env2.positions[agent]
     env2.graph.add_edge(*agents2.pos)
 
     # Test if the two graphs are identical
@@ -45,37 +49,12 @@ def test_move_agent():
 
     assert len(n1) == 1
     assert len(n2) == 0
-    assert a.pos is n1
     assert graph.positions[a] is n1
 
-    graph.move_agent(a, n2)
+    graph.move_to(a, n2)
 
     assert len(n1) == 0
     assert len(n2) == 1
-    assert a.pos is n2
-    assert graph.positions[a] is n2
-
-
-def test_move_agent_multi():
-
-    # Move agent one node to another
-    model = ap.Model()
-    graph = ap.Network(model)
-    n1 = graph.add_node()
-    n2 = graph.add_node()
-    a = ap.MultiAgent(model)
-    graph.add_agents([a], positions=[n1])
-
-    assert len(n1) == 1
-    assert len(n2) == 0
-    assert a.pos[graph] is n1
-    assert graph.positions[a] is n1
-
-    graph.move_agent(a, n2)
-
-    assert len(n1) == 0
-    assert len(n2) == 1
-    assert a.pos[graph] is n2
     assert graph.positions[a] is n2
 
 
@@ -86,7 +65,7 @@ def test_remove_agents():
     nw = ap.Network(model)
     nw.add_agents(agents)
     agent = agents[0]
-    node = agent.pos
+    node = nw.positions[agent]
     nw.remove_agents(agent)
     assert len(nw.agents) == 1
     assert len(nw.nodes) == 2
@@ -94,6 +73,6 @@ def test_remove_agents():
     assert len(nw.agents) == 1
     assert len(nw.nodes) == 1
     agent2 = agents[1]
-    nw.remove_node(agent2.pos)
+    nw.remove_node(nw.positions[agent2])
     assert len(nw.agents) == 0
     assert len(nw.nodes) == 0
