+++ agentpy/__init__.py
@@ -10,9 +10,9 @@ Source: https://github.com/JoelForamitti/agentpy
 __all__ = [
     '__version__',
     'Model',
-    'Agent', 'MultiAgent',
-    'AgentList', 'AgentGroup', 'AgentSet',
-    'AgentIter', 'AgentGroupIter', 'AttrIter',
+    'Agent',
+    'AgentList', 'AgentDList', 'AgentSet',
+    'AgentIter', 'AgentDListIter', 'AttrIter',
     'Grid', 'GridIter',
     'Space',
     'Network', 'AgentNode',
@@ -26,9 +26,9 @@ __all__ = [
 from .version import __version__
 
 from .model import Model
-from .agent import Agent, MultiAgent
-from .sequences import AgentList, AgentGroup, AgentSet
-from .sequences import AgentIter, AgentGroupIter, AttrIter
+from .agent import Agent
+from .sequences import AgentList, AgentDList, AgentSet
+from .sequences import AgentIter, AgentDListIter, AttrIter
 from .grid import Grid, GridIter
 from .space import Space
 from .network import Network, AgentNode
