### 2021-05-26 d7e2fd6
Reset record in sim_reset
