+++ docs/comparison.rst
@@ -4,27 +4,30 @@
 Comparison
 ==========
 
-Agentpy vs. Mesa
-################
+There are numerous modeling and simulation tools for ABMs,
+each with their own particular focus and style
+(find an overview `here <https://en.wikipedia.org/wiki/Comparison_of_agent-based_modeling_software>`_).
+The three main distinguishing features of agentpy are the following:
 
-An alternative framework for agent-based modeling in Python is `Mesa <https://mesa.readthedocs.io/>`__.
-The stated goal of Mesa is *"to be the Python 3-based counterpart to NetLogo, Repast, or MASON"*.
-The focus of these frameworks is traditionally on spatial environments, with an interface where one can
-observe live dynamics and adjust parameters while the model is running.
+- Agentpy integrates the multiple tasks of agent-based modeling
+  - model design, exploration with live visualization interfaces,
+  numerical experiments, and data analysis - within a single environment
+  and is optimized for interactive computing with IPython and Jupyter.
+- Agentpy is designed for scientific use with experiments over multiple runs.
+  It provides tools for parameter sampling (similar to NetLogo's BehaviorSpace),
+  Monte Carlo experiments, stochastic processes, parallel computing,
+  sensitivity analysis, and interactive visualization.
+- Agentpy is written in Python, one of the world’s most popular
+  programming languages that offers a vast number of tools and libraries for scientific use.
+  It aims to be easily compatible with established packages like
+  numpy, scipy, networkx, pandas, ema_workbench, and SALib.
 
-Agentpy, in contrast, is more focused on networks and :ref:`multi-run experiments<overview_experiments>`,
-with tools to generate and analyze :ref:`output data<overview_output>` from these experiments.
-Agentpy further has a different model structure
-that is built around :ref:`agent lists <overview_agents>`,
-which allow for simple selection and manipulation of agent groups;
-and :ref:`environments <overview_environments>`,
-which can contain agents but also act as agents themselves.
-
-To allow for an comparison of the syntax of each framework,
+The main alternative to agentpy in Python is `Mesa <https://mesa.readthedocs.io/>`__.
+To allow for an comparison of the syntax,
 here are two examples for a simple model of wealth transfer,
 both of which realize exactly the same operations.
 More information on the two models can be found in the documentation
-of each framework (link for :doc:`Agentpy <agentpy_wealth_transfer>` &
+of each framework (:doc:`Agentpy <agentpy_wealth_transfer>` &
 `Mesa <https://mesa.readthedocs.io/en/stable/tutorials/intro_tutorial.html#tutorial-description>`_).
 
 +--------------------------------------------+----------------------------------------------+
@@ -39,27 +42,23 @@ Finally, the following table provides a comparison of the main features of each 
 
 ==========================  ===================================  ===================================
 **Feature**                 **Agentpy**                          **Mesa**
-| Customizable objects      | Agent, Environment, Model	         | Agent, Model
-| Container classes         | AgentList and EnvDict for          | Scheduler (see below)
-                            | selection and manipulation
-                            | of agent and environment groups
-| Time management           | Custom activation order has to be  | Multiple scheduler classes for
-                            | defined in the Model.step method   | different activation orders
-| Supported topologies      | Spatial grid, continuous space,    | Spatial grid, continuous space,
+| Agent management          | Sequence classes                   | Scheduler classes for
+                            | like AgentList and AgentDList    | different activation orders
+| Topologies                | Spatial grid, continuous space,    | Spatial grid, continuous space,
                             | network                            | network
 | Data recording            | Recording methods for variables    | DataCollector class that can
-                            | (of agents, environments, and      | collect variables of agents
-                            | model) and evaluation measures     | and model
[…]
