+++ docs/reference_sample.rst
@@ -4,9 +4,17 @@
 Parameter samples
 =================
 
+Value sets and ranges
+#####################
+
 .. autoclass:: Range
 
+.. autoclass:: IntRange
+
 .. autoclass:: Values
+
+Sample generation
+#################
 
 .. autoclass:: Sample
     :members:
