+++ docs/guide.rst
@@ -7,6 +7,7 @@ User Guides
 This section contains interactive notebooks with common applications of the agentpy framework.
 If you are interested to add a new article to this guide, please visit :doc:`contributing`.
 If you are looking for examples of complete models, take a look at :doc:`model_library`.
+To learn how agentpy compares with other frameworks, take a look at :doc:`comparison`.
 
 .. :caption: Contents:
 .. toctree::
