+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.0.8.dev0
+version = 0.1.0.dev0
 
 [options]
 packages = find:
