+++ paper.bib
@@ -72,6 +72,14 @@
   url       = {https://mesa.readthedocs.io/},
 }
 
+@Misc{Netlogo,
+  author    = {Wilensky U},
+  title     = {NetLogo},
+  year      = {1999},
+  note      = {Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.},
+  url       = {http://ccl.northwestern.edu/netlogo/},
+}
+
 @book{North2007,
   title={Managing business complexity: discovering strategic solutions with agent-based modeling and simulation},
   author={North, Michael J and Macal, Charles M},
@@ -79,3 +87,11 @@
   publisher={Oxford University Press},
   doi={10.1093/acprof:oso/9780195172119.001.0001},
 }
+
+@misc{Wikipedia2021,
+   author = "Wikipedia",
+   title = "{Comparison of agent-based modeling software} --- {W}ikipedia{,} The Free Encyclopedia",
+   year = "2021",
+   howpublished = {\url{http://en.wikipedia.org/w/index.php?title=Comparison\%20of\%20agent-based\%20modeling\%20software&oldid=1011739575}},
+   note = "[Online; accessed 14-May-2021]"
+ }
