+++ paper.md
@@ -21,55 +21,91 @@ bibliography: paper.bib
 
 # Introduction
 
-Agentpy is an open-source library for the development and analysis of agent-based models (ABMs) in Python. 
-ABMs are computer simulations
-to generate and understand emergent dynamics of complex systems 
-based on the autonomous behavior of heterogeneous agents.
-This method is being increasingly applied in fields like
-ecology [@DeAngelis2019], cognitive sciences [@Madsen2019], management [@North2007], 
-policy analysis [@Castro2020], economics [@Farmer2009], and sociology [@Bianchi2015].
+Agent-based models are computer simulations based on the autonomous behavior of heterogeneous agents. They are used to generate and understand the emergent dynamics of complex systems, with applications in fields like ecology [@DeAngelis2019], cognitive sciences [@Madsen2019], management [@North2007], policy analysis [@Castro2020], economics [@Farmer2009], and sociology [@Bianchi2015].
 
-The aim of agentpy is to provide an intuitive syntax for the creation of models
-together with advanced tools for scientific applications.
-The framework integrates the tasks of model design, numerical experiments, 
-and subsequent data analysis and visualization within a single environment, and is
-optimized for interactive computing with [IPython](http://ipython.org/) and [Jupyter](https://jupyter.org/) (see \autoref{fig:example}). A reference of all features
-as well as a model library with tutorials and examples can be found in the [documentation](https://agentpy.readthedocs.io/).
+Agentpy is an open-source library for the development and analysis of agent-based models. It aims to provide a new intuitive syntax for the creation of models together with advanced tools for scientific applications. The framework is written in Python 3, and optimized for interactive computing with [IPython](http://ipython.org/) and [Jupyter](https://jupyter.org/). A reference of all features as well as a model library with tutorials and examples can be found in the documentation (https://agentpy.readthedocs.io/).
 
-![A screenshot of Jupyter Lab with two interactive tutorials from the agentpy model library.\label{fig:example}](docs/agentpy_example.png)
+# Statement of Need
 
-# Overview
+There are numerous modeling and simulation tools for agent-based models, each with their own particular focus and style [@Abar2017]. Most notable examples are NetLogo [@Netlogo], which is written in Scala/Java and has become the most established tool in the field; and Mesa [@Mesa2016], a more recent framework that has popularized the development of agent-based models in Python. 
 
-The framework follows a nested structure that is illustrated in \autoref{fig:structure}.
-The basic building blocks are the agents, which can be designed with custom variables and methods.
-These agents can then be placed within different environments like a network or a spatial grid.
-A model is used to initiate these objects, perform a simulation, and record data. 
-Experiments, in turn, can take a model and run it over multiple iterations and parameter combinations.
-The resulting output data can be saved, loaded, and re-arranged for analysis and visualization.
+Agentpy's main feature in comparison to these existing tools is that it integrates the many different tasks of agent-based modeling within a single environment. These tasks include the creation of custom agent, environment, and model types; interactive simulations (similar to the traditional NetLogo interface); numeric experiments over multiple runs; and subsequent data analysis of the output.
 
-![Nested structure of the agentpy framework.\label{fig:structure}](docs/structure.png){ width=70% }
[…]
