+++ paper.bib
@@ -87,11 +87,3 @@
   publisher={Oxford University Press},
   doi={10.1093/acprof:oso/9780195172119.001.0001},
 }
-
-@misc{Wikipedia2021,
-   author = "Wikipedia",
-   title = "{Comparison of agent-based modeling software} --- {W}ikipedia{,} The Free Encyclopedia",
-   year = "2021",
-   howpublished = {\url{http://en.wikipedia.org/w/index.php?title=Comparison\%20of\%20agent-based\%20modeling\%20software&oldid=1011739575}},
-   note = "[Online; accessed 14-May-2021]"
- }
