+++ paper.md
@@ -39,74 +39,6 @@ The agentpy framework follows a nested structure that is illustrated in \autoref
 
 ![Nested structure of the agentpy framework.\label{fig:structure}](docs/graphics/structure.png){ width=70% }
 
-# Usage example
-
-The following code shows an example of a simple model that explores the distribution of wealth under a randomly trading population of agents. The original version of this model has been written in Mesa [@Mesa2016], allowing for a comparison of the syntax between the two frameworks. To start, we import the agentpy library as follows:
-
-```python
-import agentpy as ap
-```
-
-We then define a new type of [`Agent`](https://agentpy.readthedocs.io/en/stable/reference_agents.html). The first method `setup` will be called automatically at the agent's creation. Each agent starts with one unit of wealth. The other method `wealth_transfer` will be called by the model during each time-step. When called, the agent randomly selects a trading partner and hands them one unit of their wealth, given that they have one to spare. 
-
-```python
-class MoneyAgent(ap.Agent):
-
-    def setup(self):
-        self.wealth = 1
-
-    def wealth_transfer(self):
-        if self.wealth > 0:
-            partner = self.model.agents.random()
-            partner.wealth += 1
-            self.wealth -= 1
-```
-
-Next, we define a [`Model`](https://agentpy.readthedocs.io/en/stable/reference_model.html). The methods `setup` and `end` will be called at the beginning and end of the simulation, while `step` is called during each time-step. An [`AgentList`](https://agentpy.readthedocs.io/en/stable/reference_sequences.html) is used to create a set of agents that can then be accessed as a group. The attribute `p` is used to access the model's parameters. And the method [`record`](https://agentpy.readthedocs.io/en/stable/reference_agents.html#agentpy.Agent.record) is used to store data for later analysis.
-
-```python
-class MoneyModel(ap.Model):
-
-    def setup(self):
-        self.agents = ap.AgentList(self, self.p.n, MoneyAgent)
-
-    def step(self):
-        self.agents.wealth_transfer()
-
-    def end(self):
-        self.agents.record('wealth')
-```
-
-To run a simulation, a new instance of the model is created with a dictionary of parameters.
-While the parameter `n` is used in the model's setup, `steps` is used automatically to define the maximum number of time-steps. Alternatively, the simulation could also be stopped with [`Model.stop`](https://agentpy.readthedocs.io/en/stable/reference_model.html#agentpy.Model.stop). To perform the actual simulation, one can use [`Model.run`](https://agentpy.readthedocs.io/en/stable/reference_model.html#agentpy.Model.run).
-
-```python
-parameters = {'n': 10, 'steps': 10}
-model = MoneyModel(parameters)
-results = model.run()
-```
-
-Parameters can also be defined as ranges and used to generate a [`Sample`](https://agentpy.readthedocs.io/en/latest/reference_sample.html).
-This sample can then be used to initiate an [`Experiment`](https://agentpy.readthedocs.io/en/latest/reference_experiment.html) that can repeatedly run the model over multiple parameter combinations and iterations. In the following example, the parameter `n` is varied from 1 to 100 and the simulation is repeated five times for each value of `n`.
-
-```python
-parameters = {'n': ap.IntRange(1, 100), 'steps': 10}
-sample = ap.Sample(parameters, n=10)
-exp = ap.Experiment(MoneyModel, sample, iterations=10, record=True)
-results = exp.run()
-```
-
[…]
