+++ paper.bib
@@ -1,89 +1,0 @@
-% Encoding: windows-1252
-
-@Article{Madsen2019,
-  author    = {Jens Koed Madsen and Richard Bailey and Ernesto Carrella and Philipp Koralus},
-  title     = {Analytic Versus Computational Cognitive Models: Agent-Based Modeling as a Tool in Cognitive Sciences},
-  journal   = {Current Directions in Psychological Science},
-  year      = {2019},
-  volume    = {28},
-  number    = {3},
-  pages     = {299-305},
-  doi       = {10.1177/0963721419834547},
-}
-
-@Article{DeAngelis2019,
-  author    = {DeAngelis, Donald L. and Diaz, Stephanie G.},
-  title     = {Decision-Making in Agent-Based Modeling: A Current Review and Future Prospectus},
-  journal   = {Frontiers in Ecology and Evolution},
-  year      = {2019},
-  volume    = {6},
-  pages     = {237},
-  doi       = {10.3389/fevo.2018.00237},
-}
-
-@Article{Farmer2009,
-  author    = {Farmer, J Doyne and Foley, Duncan},
-  title     = {The economy needs agent-based modelling},
-  journal   = {Nature},
-  year      = {2009},
-  volume    = {460},
-  number    = {7256},
-  pages     = {685--686},
-  doi       = {10.1038/460685a},
-}
-
-@Article{Castro2020,
-  author    = {Castro, Juana and Drews, Stefan and Exadaktylos, Filippos and Foramitti, Joël and Klein, Franziska and Konc, Théo and Savin, Ivan and van den Bergh, Jeroen},
-  title     = {A review of agent-based modeling of climate-energy policy},
-  journal   = {WIREs Climate Change},
-  year      = {2020},
-  volume    = {11},
-  number    = {4},
-  pages     = {e647},
-  doi       = {10.1002/wcc.647},
-}
-
-@Article{Bianchi2015,
-  author    = {Bianchi, Federico and Squazzoni, Flaminio},
-  title     = {Agent-based models in sociology},
-  journal   = {WIREs Computational Statistics},
-  year      = {2015},
-  volume    = {7},
-  number    = {4},
-  pages     = {284-306},
-  doi       = {10.1002/wics.1356},
-}
-
-@Article{Abar2017,
-  author    = {Sameera Abar and Georgios K. Theodoropoulos and Pierre Lemarinier and Gregory M.P. O’Hare},
-  title     = {Agent Based Modelling and Simulation tools: A review of the state-of-art software},
-  journal   = {Computer Science Review},
-  year      = {2017},
-  volume    = {24},
-  pages     = {13 - 33},
-  doi       = {10.1016/j.cosrev.2017.03.001},
-}
-
-@Misc{Mesa2016,
-  author    = {{Project Mesa Team}},
-  title     = {Mesa: Agent-based modeling in Python 3+},
-  year      = {2016},
-  note      = {Accessed: 2020-12-29},
-  url       = {https://mesa.readthedocs.io/},
-}
-
-@Misc{Netlogo,
-  author    = {Wilensky U},
-  title     = {NetLogo},
-  year      = {1999},
-  note      = {Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.},
-  url       = {http://ccl.northwestern.edu/netlogo/},
-}
-
-@book{North2007,
-  title={Managing business complexity: discovering strategic solutions with agent-based modeling and simulation},
-  author={North, Michael J and Macal, Charles M},
-  year={2007},
-  publisher={Oxford University Press},
-  doi={10.1093/acprof:oso/9780195172119.001.0001},
-}
