+++ tests/test_experiment.py
@@ -64,7 +64,7 @@ def test_random():
             self.report('x', self.model.random.random())
 
     sample = ap.Sample(parameters)
-    exp = ap.Experiment(Model, sample, iterations=2, random=True)
+    exp = ap.Experiment(Model, sample, iterations=2, randomize=True)
     results = exp.run()
 
     l = list(results.reporters['x'])
