+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Agentpy: A package for agent-based modeling in Python'
+title: 'AgentPy: A package for agent-based modeling in Python'
 tags:
   - Agent-based modeling
   - Complex systems
@@ -18,6 +18,99 @@ affiliations:
 date: 16.01.2020
 bibliography: paper.bib
 ---
+
+# Introduction
+
+Agent-based models allow for computer simulations based on the autonomous behavior of heterogeneous agents. They are used to generate and understand the emergent dynamics of complex systems, with applications in fields like ecology [@DeAngelis2019], cognitive sciences [@Madsen2019], management [@North2007], policy analysis [@Castro2020], economics [@Arthur2021; @Farmer2009], and sociology [@Bianchi2015].
+
+AgentPy is an open-source library for the development and analysis of agent-based models. It aims to provide an intuitive syntax for the creation of models together with advanced tools for scientific applications. The framework is written in Python 3, and optimized for interactive computing with [IPython](http://ipython.org/) and [Jupyter](https://jupyter.org/). A reference of all features as well as a model library with tutorials and examples can be found in the [documentation](https://agentpy.readthedocs.io/).[^1]
+
+# Statement of Need
+
+There are numerous modeling and simulation tools for agent-based models, each with their own particular focus and style [@Abar2017]. Notable examples are NetLogo [@Netlogo], which is written in Scala/Java and has become the most established tool in the field; and [Mesa](https://mesa.readthedocs.io/), a more recent framework that has popularized the development of agent-based models in Python.[^2]
+
+A main distinguishing feature of AgentPy is that it integrates the many different tasks of agent-based modeling within a single environment for interactive computing. This includes the creation of custom agent and model types; interactive simulations (similar to the traditional NetLogo interface); numeric experiments over multiple runs; and the subsequent data analysis of the output. All of these can be performed within a [Jupyter Notebook](https://jupyter.org/).
+
+The software is further designed for scientific applications, and includes tools for parameter sampling (similar to NetLogo's BehaviorSpace), Monte Carlo experiments, random number generation, parallel computing, and sensitivity analysis. Beyond these built-in features, AgentPy is also designed for compatibility with established Python packages like [EMA Workbench](https://emaworkbench.readthedocs.io/), [NetworkX](https://networkx.org/), [NumPy](https://numpy.org/), [pandas](https://pandas.pydata.org/), [SALib](https://salib.readthedocs.io/), [SciPy](https://www.scipy.org/), and [seaborn](https://seaborn.pydata.org/).
+
+# Basic structure
+
+The AgentPy framework follows a nested structure that is illustrated in \autoref{fig:structure}. The basic building blocks are the agents, which can be placed within (multiple) environments with different topologies such as a network, a spatial grid, or a continuous space. Models are used to initiate these objects, perform a simulation, and record data. Experiments can run a model over multiple iterations and parameter combinations. The resulting output data can then be saved and re-arranged for analysis and visualization.
+
+![Nested structure of the AgentPy framework.\label{fig:structure}](docs/graphics/structure.png){ width=60% }
+
+# Model example
+
+The following code shows an example of a simple model that explores the distribution of wealth under a randomly trading population of agents. The original version of this model was written in Mesa, allowing for a comparison of the syntax between the two frameworks.[^3] To start, we import the AgentPy library as follows:
+
+```python
+import agentpy as ap
+```
+
[…]
