+++ README.md
@@ -1,4 +1,4 @@
-# Agentpy - Agent-based modeling in Python
+# AgentPy - Agent-based modeling in Python
 
 [![PyPI](https://img.shields.io/pypi/v/agentpy)](https://pypi.org/project/agentpy/)
 [![GitHub](https://img.shields.io/github/license/joelforamitti/agentpy)](https://github.com/JoelForamitti/agentpy/blob/master/LICENSE)
@@ -6,8 +6,8 @@
 [![Documentation Status](https://readthedocs.org/projects/agentpy/badge/?version=latest)](https://agentpy.readthedocs.io/en/latest/?badge=latest)
 [![codecov](https://codecov.io/gh/JoelForamitti/agentpy/branch/master/graph/badge.svg?token=NTW99HNGB0)](https://codecov.io/gh/JoelForamitti/agentpy)
 
-Agentpy is an open-source library for the development and analysis of agent-based models in Python. 
-The framework integrates the tasks of model design, numerical experiments, 
+AgentPy is an open-source library for the development and analysis of agent-based models in Python. 
+The framework integrates the tasks of model design, interactive simulations, numerical experiments, 
 and data analysis within a single environment, and is optimized for interactive computing 
 with [IPython](http://ipython.org/) and [Jupyter](https://jupyter.org/). 
 
