+++ docs/conf.py
@@ -12,8 +12,6 @@ from agentpy import __version__  # Agentpy must be installed first
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 
-sys.path.insert(0, os.path.abspath('../examples'))
-
 # -- Project information -----------------------------------------------------
 
 project = 'agentpy'
