### 2021-05-28 JoelForamitti/agentpy#13 @JoelForamitti
Dear @jamesdamillington

Thank you very much for the useful comments. I have just released AgentPy 0.1.0, together with a revised manuscript. Below I address your comments.

> One minor comment is that in the contents of the API reference, I think it might be useful to add (Grid) to the title entry for Discrete spaces (to make the entry Discrete spaces (Grid) and even clearer the link to how this is described in the Overview - see comment about environments below).

This has been changed as suggested.

> Throughout the docs, the word separate is frequently misspelled as seperate (also in the Flocking example model) - it's a minor issue but as you edit documentation in future you might fix this.

This has been corrected throughout the documentation.

> The Using Environments section is a little confusing as it states, "There are three different types of environments:" then presents four items in the list following. I think this section could clarify better the relationship of Environment to the other types.

To avoid confusion, there are now only three types: Grid, Space, and Network.

> For example, you might also highlight here that the Wealth Transfer example model does not use any of Network, Grid, or Space (it just assumes an env with no topology).

This is now mentioned in the environment section of overview.

> Also, you might state explicitly here that multiple types of environment can operate simultaneously (I infer this is true from the documentation on Agent.neighbours which states "If an agent has the same neighbors in multiple environments, ..."

To make this clearer, the documentation now states: “A model can contain zero, one or multiple environments which agents can enter and leave.”

> The Recording data section: it might be useful in this section to emphasise (even more clearly) that this is about recording the model state for output to file and later analysis. I initially mistakenly thought it referred to what information might be accessed at runtime (e.g. I wonder why agents would not be able to access aggregate level information about the state of the system dynamically, but the model examples show this is clearly possible). Once I understood the intention of agentpy is to be used for experiments with multiple runs the importance of this section became clear (but see my comments below on making that intention more prominent). One way to make this clearer might be to refer to the Output and analysis section and make clear how record() and Model.measure() are related to DataDict.

Please note that the name ‘measures’ has been changed to ‘reporters’ in the latest update.
It is now stated explicitly that the purpose of recording & reporting is for later analysis.
However, recorded data can still be accessed at run-time through the `log` and `reporters` attributes, which is now also mentioned in the overview section on recording data.

> The comparison with Mesa is very useful, both in the paper and in the documentation. I think you should emphasise this more, to make clear why people might want to use agentpy:
> ·	in the paper, highlighting that agentpy has been designed for scientific use with experiments from multiple runs would be useful in the Introduction, rather than leaving it to the final paragraph of the paper

 This is now explicitly mentioned in the new Statement of Need section.

> ·	in the documentation, I suggest you link to the Comparison page from the User Guide page (and emphasise what sorts of modelling purposes you have in mind for agentpy on the User Guide page also)

The user guide now links to comparison, and the comparison page has a more explicit explanation of the frameworks distinguishing features/aims.

> I wonder if in future you could also provide a comparison to NetLogo.  NetLogo is a very useful starting point for people learning agent-based  simulation. To attract people from NetLogo to use agentpy a comparison  would be useful I think. Building on this, some brief thoughts from  someone (myself) coming at this with a background in using NetLogo:
> ·	make clear your Grid and Space environments are not toroidal (which is the default case in NetLogo)

With the new update, both grid and space have the option to be toroidal.

> ·	highlight that experiments are akin to NetLogo's behaviour space

This is now mentioned in both the paper and the documentation.

> ·	comment on how patches and patchsets (NetLogo's grid) would be implemented in agentpy

This is something I am still working on. I will try to add a comparison for this in the future.

> I cannot see any reference currently to automated tests in the  documentation. I can see that such tests might be possible given the  stochastic nature of ABM. You might recommend users run the Example  Models to check their output is generally the same as that in the  examples shown on the docs webpages. Please comment

All functionalities of the library have been tested. There is a shield on the landing page of the repository and documentation that leads to a report on the the code coverage of these tests. See https://app.codecov.io/gh/JoelForamitti/agentpy

> However, one addition to the contributions page might be to welcome contributions of (simple) models to the Model Library to expand the examples available?

This has been added as follows: “Contributions of clear and simple demonstration models for the model library that illustrate a particular application are also very welcome.”

> Statement of need
> The comparison to other packages is present and useful. However,  while the Abar et al. (2017) paper is good, it is not open access. Could  you either:
> ·	add an open-access equivalent review paper, or
> ·	add links to webpages for the frameworks mentioned (NetLogo, Repast, MASON)

The paper now only mentions NetLogo, for which I added a reference, and Mesa, for which I added an explicit link to the documentation.
