### 2021-05-28 JoelForamitti/agentpy#16 @JoelForamitti
Dear @martibosch

Thank you very much for your helpful review. I have just released AgentPy 0.1.0, together with a revised manuscript. Below I address your comments.

> Statement of need
> AFAIK, this section must be explicit in the manuscript. I suppose that such a section would include the sentence of lines 13-14. I would further suggest to make it a "self-contained" sentence (that does not require reading any other part of the manuscript to fully understand it), hence I would add:
> The aim of agentpy is to provide an intuitive syntax for the creation of **agent-based** models together with advanced tools for scientific applications.

There is now an explicit Statement of Need section in the manuscript. Although a different sentence is used, the term 'agent-based model' has been written out.

> Figure and code snippets
> ·	Figure 1: it may be better to use a code snippet rather than a figure to display code, e.g., users might want to copy-paste them
> ·	lines 27-30: it is better that the text refers to some code snippet in the manuscript (e.g., Figure 1). For instance: from the current manuscript, one may ask: what is agents? I had to open a notebook, instantiate some example agentpy.Model, inspect the class via dir and it is only then that I found the agents attribute

Thank you for pointing this out. Figure 1 has been removed and replaced by a whole new section ‘Usage Example’ that includes multiple code snippets.

The newest update of agentpy has also improved the syntax, so that the `agent` attribute now has to be declared explicitly in the model definition, as can be seen in the snippets.

> Clarify some concepts and definitions
> ·	line 30: what is the "record" action? I needed to go to the docstring of the record method to fully understand what it means. It may be a good idea to add a brief sentence to clarify it within the manuscript so that the reader can get a clearer picture there.

Apart from now providing full code snippets, a short sentence explaining record has been added and if one clicks on record they will be linked to the documentation.

> ·	line 45: "direct integration of analysis tools within the same framework" as somebody familiar with scientific Python, I suppose that you mean NumPy, pandas, NetworkX and the like, but it may be better to make it more explicit.

All the supported packages are now explicitly mentioned in the paper. They include EMA Workbench, NetworkX, NumPy, pandas, SALib, SciPy, and seaborn.

> Other
> ·	line 29: IMHO "attribute" seems more Pythonic than "variable"

This line has been removed in the revised paper.

> ·	links: it might be better to put URLs explicitly (e.g., although unlikely in JOSS, people might print the paper). Line 18 is fine since we can easily find the documentation without clicking the link, but not so much for line 46, where I believe that the word "here" should be replaced by the actual URL.

Explicit links are now given in footnotes for the AgentPy documentation and discussion forum, the comparison page, and the Mesa documentation.
