### 2021-05-28 #3065 @JoelForamitti
Dear @sbenthall, @jamesdamillington, & @martibosch,

above you can find the revised manuscript for my submission. it comes together with a new version of the software (v0.1.0) that has been released today. i have tried to take all the comments from the review into account, and also added some new parts to include the new features that have been released since the original submission in January.
