+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.1.0
+version = 0.1.1.dev
 
 [options]
 packages = find:
