+++ docs/guide.rst
@@ -1,5 +1,3 @@
-.. currentmodule:: agentpy
-
 ===========
 User Guides
 ===========
@@ -14,7 +12,6 @@ To learn how agentpy compares with other frameworks, take a look at :doc:`compar
    :caption: Contents
    :maxdepth: 1
 
-    guide_interactive
-    guide_random
-    guide_ema
-
+   guide_interactive
+   guide_random
+   guide_ema
