### 2021-05-31 JoelForamitti/agentpy#16 @martibosch
Hello again @JoelForamitti!

I have revised your comments and I find that most of my remarks have been addressed. The only part that I did not understand is the `agent` attribute, i.e.:

> The newest update of agentpy has also improved the syntax, so that the `agent` attribute now has to be declared explicitly in the model definition, as can be seen in the snippets.

what is the `agent` attribute? I cannot find it in any of the snippets - did you mean the `agents` attribute? If so, it is not declared explicitly in the `WealthAgent` snippet.

Best,
Martí
