### 2021-05-31 JoelForamitti/agentpy#16 @JoelForamitti
Hi @martibosch

that was a typo from my side, i indeed meant `agents`, sorry.

I copy here the model definition from the paper. You can see that `agents` is being declared in the model's setup function:

```python
class WealthModel(ap.Model):

    def setup(self):
        self.agents = ap.AgentList(self, self.p.n, WealthAgent)

    def step(self):
        self.agents.wealth_transfer()

    def end(self):
        self.agents.record('wealth')
```

In the `WealthAgent` snippet, the same attribute is accessed through `model.agents`.

Best,
Joël
