### 2021-05-31 JoelForamitti/agentpy#16 @martibosch
> that was a typo from my side, i indeed meant `agents`, sorry.

this is what I imagined, thank you for clarifying. However, [the latest (if I am not mistaken) version of the manuscript](https://github.com/openjournals/joss-reviews/issues/3065#issuecomment-850435847)  shows a different snippet (between lines 48 and 49) which does not explicitly declare the `agents` attribute in the `setup` method. Am I missing something?

Best,
Martí
