### 2021-05-31 JoelForamitti/agentpy#16 @JoelForamitti
The `agents` are declared not in the agent (lines 48-49), but in the model (lines 53-54, which is the snippet that i copied here).
