### 2021-05-31 JoelForamitti/agentpy#16 @martibosch
Indeed, my mistake, thank you for pointing it out!

Closing this then.
