### 2021-06-04 #3065 @martibosch
I just have one last question regarding the "Summary" checkbox. I believe that the second paragraph of the "Introduction" of the current manuscript provides a concise summary of agentpy, however, do JOSS papers require a dedicated "Summary" section? If not, I can check off the box and the manuscript can proceed to publication.
