### 2021-06-07 JoelForamitti/agentpy#13 @jamesdamillington
@JoelForamitti Thanks for your detailed responses that address my comments well. I think the revised manuscript and new version look good and I will now close this issue.
