### 2021-06-07 #3065 @jamesdamillington
The revisions look good to me (and I have closed [the issue](https://github.com/JoelForamitti/agentpy/issues/13) I opened). I did [wonder the same things](https://github.com/openjournals/joss-reviews/issues/3065#issuecomment-854723095) as @martibosch about the 'summary', but I also agree that the second paragraph of the Introduction section is a good summary if a dedicated Summary section is not needed (hence I have checked the Summary checklist item). I am happy that this can move to publication.
