+++ paper/paper.bib
@@ -90,3 +90,13 @@
   publisher={Nature Publishing Group},
   doi={10.1038/s42254-020-00273-3}
 }
+
+@InProceedings{ Mesa2015,
+  author    = { {D}avid {M}asad and {J}acqueline {K}azil },
+  title     = { {M}esa: {A}n {A}gent-{B}ased {M}odeling {F}ramework },
+  booktitle = { {P}roceedings of the 14th {P}ython in {S}cience {C}onference },
+  pages     = { 51 - 58 },
+  year      = { 2015 },
+  editor    = { {K}athryn {H}uff and {J}ames {B}ergstra },
+  doi       = { 10.25080/Majora-7b98e3ed-009 }
+}
