### 2021-06-09 #3065 @xuanxu
> Is it necessary to have a dedicated "Summary" section, or is having the summary in the introduction sufficient?

@sbenthall The paper should include a summary, but is not mandatory to title the section as 'Summary', it is recommended, but 'Introduction' is also OK.
