### 2021-06-14 #3065 @sbenthall
I've proofread the article. It looks great. I'm also excited for the publication of this work -- this is software I could use in my work!

@JoelForamitti could you please make a new tagged release and archive it using Zenoda, and then report the version number and archive DOI in this thread?
