### 2021-06-14 bd726f2
[tests] Grid.move_to
