+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.1.1.dev
+version = 0.1.1
 
 [options]
 packages = find:
