### 2021-06-14 #3065 @JoelForamitti
@sbenthall thank you for your kind words!
here is the new release:

Version: v0.1.1
DOI: 10.5281/zenodo.4946508
Link: https://doi.org/10.5281/zenodo.4946508
