+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.1.1
+version = 0.1.2.dev
 
 [options]
 packages = find:
