### 2021-06-15 652f9e7
[docs] Requirements for syntax highlighting
