+++ setup.cfg
@@ -20,6 +20,7 @@ docs =
     sphinx >= 3.2.1
     nbsphinx >= 0.7.1
     sphinx_rtd_theme >= 0.4.3
+    IPython >= 7.19.0
 dev =
     setuptools >= 51.3.3
     pytest >= 5.4.3
