+++ docs/about.rst
@@ -8,13 +8,23 @@ Agentpy has been created by Joël Foramitti and is
 available under the open-source `BSD 3-Clause <https://github.com/JoelForamitti/agentpy/blob/master/LICENSE>`_ license.
 Source files can be found on the `GitHub repository <https://github.com/joelforamitti/agentpy>`_.
 
-Parts of this package where created with Cookiecutter_
-and the `audreyr/cookiecutter-pypackage`_ project template.
+Thanks to everyone who has contributed
+or supported the developement of this package:
 
-This project has further benefited from an ERC Advanced Grant
+- Jeroen C.J.M. van den Bergh
+- Ivan Savin
+- James Millington
+- Martí Bosch
+- Sebastian Benthall
+- Bhakti Stephan Onggo
+
+This project has benefited from an ERC Advanced Grant
 from the European Research Council (ERC)
 under the European Union's Horizon 2020 research and innovation programme
 (grant agreement n° 741087).
+
+Parts of this package where created with Cookiecutter_
+and the `audreyr/cookiecutter-pypackage`_ project template.
 
 .. _Cookiecutter: https://github.com/audreyr/cookiecutter
 .. _`audreyr/cookiecutter-pypackage`: https://github.com/audreyr/cookiecutter-pypackage
