### 2021-06-15 4813a18
[new] AgentIter.to_list & Network.nodes
