+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.1.2.dev
+version = 0.1.2
 
 [options]
 packages = find:
