### 2021-06-24 #3065 @JoelForamitti
Hi @kyleniemeyer, I realized there is typo in the acknowledgements! Is there a way to still fix this?
