### 2021-06-24 #3065 @arfon
@JoelForamitti – that should be updated now. The updated paper might take a few hours to show up properly on the JOSS site due to caching.
